% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anndataUtils.R
\name{AnndataUtils}
\alias{AnndataUtils}
\title{Utils for Anndata}
\description{
Utils for Anndata
}
\examples{
TODO
}
\section{Super classes}{
\code{\link[ceceliaCore:ReactiveObject]{ceceliaCore::ReactiveObject}} -> \code{\link[ceceliaCore:PopulationUtils]{ceceliaCore::PopulationUtils}} -> \code{AnndataUtils}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AnndataUtils-new}{\code{AnndataUtils$new()}}
\item \href{#method-AnndataUtils-popDT}{\code{AnndataUtils$popDT()}}
\item \href{#method-AnndataUtils-adataMat}{\code{AnndataUtils$adataMat()}}
\item \href{#method-AnndataUtils-save}{\code{AnndataUtils$save()}}
\item \href{#method-AnndataUtils-popPaths}{\code{AnndataUtils$popPaths()}}
\item \href{#method-AnndataUtils-isLabelPropsStore}{\code{AnndataUtils$isLabelPropsStore()}}
\item \href{#method-AnndataUtils-getAdataDT}{\code{AnndataUtils$getAdataDT()}}
\item \href{#method-AnndataUtils-clone}{\code{AnndataUtils$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="ReactiveObject" data-id="reactive"><a href='../../ceceliaCore/html/ReactiveObject.html#method-ReactiveObject-reactive'><code>ceceliaCore::ReactiveObject$reactive()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="ReactiveObject" data-id="update"><a href='../../ceceliaCore/html/ReactiveObject.html#method-ReactiveObject-update'><code>ceceliaCore::ReactiveObject$update()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="PopulationUtils" data-id="defaultPops"><a href='../../ceceliaCore/html/PopulationUtils.html#method-PopulationUtils-defaultPops'><code>ceceliaCore::PopulationUtils$defaultPops()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="PopulationUtils" data-id="getImChannelLimits"><a href='../../ceceliaCore/html/PopulationUtils.html#method-PopulationUtils-getImChannelLimits'><code>ceceliaCore::PopulationUtils$getImChannelLimits()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="PopulationUtils" data-id="getImChannels"><a href='../../ceceliaCore/html/PopulationUtils.html#method-PopulationUtils-getImChannels'><code>ceceliaCore::PopulationUtils$getImChannels()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="PopulationUtils" data-id="getPopObj"><a href='../../ceceliaCore/html/PopulationUtils.html#method-PopulationUtils-getPopObj'><code>ceceliaCore::PopulationUtils$getPopObj()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="PopulationUtils" data-id="getPopObjFilepath"><a href='../../ceceliaCore/html/PopulationUtils.html#method-PopulationUtils-getPopObjFilepath'><code>ceceliaCore::PopulationUtils$getPopObjFilepath()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="PopulationUtils" data-id="getValueNames"><a href='../../ceceliaCore/html/PopulationUtils.html#method-PopulationUtils-getValueNames'><code>ceceliaCore::PopulationUtils$getValueNames()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="PopulationUtils" data-id="popLabels"><a href='../../ceceliaCore/html/PopulationUtils.html#method-PopulationUtils-popLabels'><code>ceceliaCore::PopulationUtils$popLabels()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="PopulationUtils" data-id="savePops"><a href='../../ceceliaCore/html/PopulationUtils.html#method-PopulationUtils-savePops'><code>ceceliaCore::PopulationUtils$savePops()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="PopulationUtils" data-id="unlinkPops"><a href='../../ceceliaCore/html/PopulationUtils.html#method-PopulationUtils-unlinkPops'><code>ceceliaCore::PopulationUtils$unlinkPops()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnndataUtils-new"></a>}}
\if{latex}{\out{\hypertarget{method-AnndataUtils-new}{}}}
\subsection{Method \code{new()}}{
init
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnndataUtils$new(objFilepath, imChannels)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{objFilepath}}{character of object file path}

\item{\code{imChannels}}{list of character for channel names}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnndataUtils-popDT"></a>}}
\if{latex}{\out{\hypertarget{method-AnndataUtils-popDT}{}}}
\subsection{Method \code{popDT()}}{
Population data.table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnndataUtils$popDT(
  pops = NULL,
  cols = NULL,
  dropNA = FALSE,
  dropPop = FALSE,
  copyDT = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pops}}{list of character for populations}

\item{\code{cols}}{list of character for columns}

\item{\code{dropNA}}{boolean to drop NA}

\item{\code{dropPop}}{boolean to drop population}

\item{\code{copyDT}}{boolean to copy data.table}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnndataUtils-adataMat"></a>}}
\if{latex}{\out{\hypertarget{method-AnndataUtils-adataMat}{}}}
\subsection{Method \code{adataMat()}}{
Channel matrix
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnndataUtils$adataMat()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnndataUtils-save"></a>}}
\if{latex}{\out{\hypertarget{method-AnndataUtils-save}{}}}
\subsection{Method \code{save()}}{
Save anndata back to disk
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnndataUtils$save()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnndataUtils-popPaths"></a>}}
\if{latex}{\out{\hypertarget{method-AnndataUtils-popPaths}{}}}
\subsection{Method \code{popPaths()}}{
Populations paths
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnndataUtils$popPaths(includeRoot = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{includeRoot}}{boolean to include root}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnndataUtils-isLabelPropsStore"></a>}}
\if{latex}{\out{\hypertarget{method-AnndataUtils-isLabelPropsStore}{}}}
\subsection{Method \code{isLabelPropsStore()}}{
Whether utils are based on label props
for clusters, pull information from the main DT
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnndataUtils$isLabelPropsStore()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnndataUtils-getAdataDT"></a>}}
\if{latex}{\out{\hypertarget{method-AnndataUtils-getAdataDT}{}}}
\subsection{Method \code{getAdataDT()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnndataUtils$getAdataDT()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnndataUtils-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AnndataUtils-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnndataUtils$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
