% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cciaImage.R
\name{CciaImage}
\alias{CciaImage}
\title{Reactive image object}
\description{
Reactive image object
}
\examples{
TODO
}
\section{Super classes}{
\code{\link[cecelia:ReactiveObject]{cecelia::ReactiveObject}} -> \code{\link[cecelia:ReactivePersistentObject]{cecelia::ReactivePersistentObject}} -> \code{CciaImage}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CciaImage-new}{\code{CciaImage$new()}}
\item \href{#method-CciaImage-summary}{\code{CciaImage$summary()}}
\item \href{#method-CciaImage-editChannelName}{\code{CciaImage$editChannelName()}}
\item \href{#method-CciaImage-omeXML}{\code{CciaImage$omeXML()}}
\item \href{#method-CciaImage-omeXMLPath}{\code{CciaImage$omeXMLPath()}}
\item \href{#method-CciaImage-omeXMLPixels}{\code{CciaImage$omeXMLPixels()}}
\item \href{#method-CciaImage-omeXMLTimelapseInfo}{\code{CciaImage$omeXMLTimelapseInfo()}}
\item \href{#method-CciaImage-omeXMLPixelRes}{\code{CciaImage$omeXMLPixelRes()}}
\item \href{#method-CciaImage-resetMetaFromFile}{\code{CciaImage$resetMetaFromFile()}}
\item \href{#method-CciaImage-tracksInfo}{\code{CciaImage$tracksInfo()}}
\item \href{#method-CciaImage-tracksMeasures}{\code{CciaImage$tracksMeasures()}}
\item \href{#method-CciaImage-tracks}{\code{CciaImage$tracks()}}
\item \href{#method-CciaImage-popStats}{\code{CciaImage$popStats()}}
\item \href{#method-CciaImage-labelPropsCols}{\code{CciaImage$labelPropsCols()}}
\item \href{#method-CciaImage-labelProps}{\code{CciaImage$labelProps()}}
\item \href{#method-CciaImage-regionProps}{\code{CciaImage$regionProps()}}
\item \href{#method-CciaImage-neighbourProps}{\code{CciaImage$neighbourProps()}}
\item \href{#method-CciaImage-popUtils}{\code{CciaImage$popUtils()}}
\item \href{#method-CciaImage-popPathsAll}{\code{CciaImage$popPathsAll()}}
\item \href{#method-CciaImage-popPaths}{\code{CciaImage$popPaths()}}
\item \href{#method-CciaImage-neighboursDT}{\code{CciaImage$neighboursDT()}}
\item \href{#method-CciaImage-popDT}{\code{CciaImage$popDT()}}
\item \href{#method-CciaImage-savePops}{\code{CciaImage$savePops()}}
\item \href{#method-CciaImage-pp3}{\code{CciaImage$pp3()}}
\item \href{#method-CciaImage-ppp}{\code{CciaImage$ppp()}}
\item \href{#method-CciaImage-spatialDT}{\code{CciaImage$spatialDT()}}
\item \href{#method-CciaImage-spatialGraph}{\code{CciaImage$spatialGraph()}}
\item \href{#method-CciaImage-flowFrame}{\code{CciaImage$flowFrame()}}
\item \href{#method-CciaImage-flowGatingSet}{\code{CciaImage$flowGatingSet()}}
\item \href{#method-CciaImage-clsfPopUtils}{\code{CciaImage$clsfPopUtils()}}
\item \href{#method-CciaImage-livePopUtils}{\code{CciaImage$livePopUtils()}}
\item \href{#method-CciaImage-adataUtils}{\code{CciaImage$adataUtils()}}
\item \href{#method-CciaImage-savePopMap}{\code{CciaImage$savePopMap()}}
\item \href{#method-CciaImage-addFilteredPops}{\code{CciaImage$addFilteredPops()}}
\item \href{#method-CciaImage-addPop}{\code{CciaImage$addPop()}}
\item \href{#method-CciaImage-setPopAttr}{\code{CciaImage$setPopAttr()}}
\item \href{#method-CciaImage-popAttr}{\code{CciaImage$popAttr()}}
\item \href{#method-CciaImage-popIDFromPath}{\code{CciaImage$popIDFromPath()}}
\item \href{#method-CciaImage-popIDsFromPaths}{\code{CciaImage$popIDsFromPaths()}}
\item \href{#method-CciaImage-popLeaves}{\code{CciaImage$popLeaves()}}
\item \href{#method-CciaImage-popIDsByAttrList}{\code{CciaImage$popIDsByAttrList()}}
\item \href{#method-CciaImage-popIDsByAttr}{\code{CciaImage$popIDsByAttr()}}
\item \href{#method-CciaImage-popIDByAttr}{\code{CciaImage$popIDByAttr()}}
\item \href{#method-CciaImage-delPop}{\code{CciaImage$delPop()}}
\item \href{#method-CciaImage-delPopsByPath}{\code{CciaImage$delPopsByPath()}}
\item \href{#method-CciaImage-toggleVisibilityForPop}{\code{CciaImage$toggleVisibilityForPop()}}
\item \href{#method-CciaImage-editPopName}{\code{CciaImage$editPopName()}}
\item \href{#method-CciaImage-editPopColour}{\code{CciaImage$editPopColour()}}
\item \href{#method-CciaImage-saveData}{\code{CciaImage$saveData()}}
\item \href{#method-CciaImage-savePopsToLabels}{\code{CciaImage$savePopsToLabels()}}
\item \href{#method-CciaImage-resetData}{\code{CciaImage$resetData()}}
\item \href{#method-CciaImage-loadData}{\code{CciaImage$loadData()}}
\item \href{#method-CciaImage-values}{\code{CciaImage$values()}}
\item \href{#method-CciaImage-valueDefault}{\code{CciaImage$valueDefault()}}
\item \href{#method-CciaImage-valueNamesForPopType}{\code{CciaImage$valueNamesForPopType()}}
\item \href{#method-CciaImage-valueNames}{\code{CciaImage$valueNames()}}
\item \href{#method-CciaImage-valuePartOf}{\code{CciaImage$valuePartOf()}}
\item \href{#method-CciaImage-valueSuffixes}{\code{CciaImage$valueSuffixes()}}
\item \href{#method-CciaImage-valueLocUID}{\code{CciaImage$valueLocUID()}}
\item \href{#method-CciaImage-valueAttr}{\code{CciaImage$valueAttr()}}
\item \href{#method-CciaImage-setOriFilepath}{\code{CciaImage$setOriFilepath()}}
\item \href{#method-CciaImage-setImFilepath}{\code{CciaImage$setImFilepath()}}
\item \href{#method-CciaImage-setImMeta}{\code{CciaImage$setImMeta()}}
\item \href{#method-CciaImage-setImChannelNames}{\code{CciaImage$setImChannelNames()}}
\item \href{#method-CciaImage-setImLabelsFilepath}{\code{CciaImage$setImLabelsFilepath()}}
\item \href{#method-CciaImage-setImLabelPropsFilepath}{\code{CciaImage$setImLabelPropsFilepath()}}
\item \href{#method-CciaImage-setImRegionsFilepath}{\code{CciaImage$setImRegionsFilepath()}}
\item \href{#method-CciaImage-setImNeighboursFilepath}{\code{CciaImage$setImNeighboursFilepath()}}
\item \href{#method-CciaImage-setImPopMap}{\code{CciaImage$setImPopMap()}}
\item \href{#method-CciaImage-setImPopGatePlot}{\code{CciaImage$setImPopGatePlot()}}
\item \href{#method-CciaImage-setImGatingSetFilepath}{\code{CciaImage$setImGatingSetFilepath()}}
\item \href{#method-CciaImage-setImAnndataFilepath}{\code{CciaImage$setImAnndataFilepath()}}
\item \href{#method-CciaImage-imMeta}{\code{CciaImage$imMeta()}}
\item \href{#method-CciaImage-imChannelNames}{\code{CciaImage$imChannelNames()}}
\item \href{#method-CciaImage-oriFilepath}{\code{CciaImage$oriFilepath()}}
\item \href{#method-CciaImage-imFilepath}{\code{CciaImage$imFilepath()}}
\item \href{#method-CciaImage-imLabelsFilepath}{\code{CciaImage$imLabelsFilepath()}}
\item \href{#method-CciaImage-imRegionsFilepath}{\code{CciaImage$imRegionsFilepath()}}
\item \href{#method-CciaImage-imNeighboursFilepath}{\code{CciaImage$imNeighboursFilepath()}}
\item \href{#method-CciaImage-imLabelPropsFilepath}{\code{CciaImage$imLabelPropsFilepath()}}
\item \href{#method-CciaImage-imGatingSetFilepath}{\code{CciaImage$imGatingSetFilepath()}}
\item \href{#method-CciaImage-imAnndataFilepath}{\code{CciaImage$imAnndataFilepath()}}
\item \href{#method-CciaImage-imShapes}{\code{CciaImage$imShapes()}}
\item \href{#method-CciaImage-imPopMap}{\code{CciaImage$imPopMap()}}
\item \href{#method-CciaImage-imPopGatePlots}{\code{CciaImage$imPopGatePlots()}}
\item \href{#method-CciaImage-clone}{\code{CciaImage$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactiveObject" data-id="reactive"><a href='../../cecelia/html/ReactiveObject.html#method-ReactiveObject-reactive'><code>cecelia::ReactiveObject$reactive()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactiveObject" data-id="update"><a href='../../cecelia/html/ReactiveObject.html#method-ReactiveObject-update'><code>cecelia::ReactiveObject$update()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="addCciaAttr"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-addCciaAttr'><code>cecelia::ReactivePersistentObject$addCciaAttr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="delCciaAttr"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-delCciaAttr'><code>cecelia::ReactivePersistentObject$delCciaAttr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="deleteObjectDirectory"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-deleteObjectDirectory'><code>cecelia::ReactivePersistentObject$deleteObjectDirectory()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="editCciaAttr"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-editCciaAttr'><code>cecelia::ReactivePersistentObject$editCciaAttr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="finalize"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-finalize'><code>cecelia::ReactivePersistentObject$finalize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="getCciaAttr"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-getCciaAttr'><code>cecelia::ReactivePersistentObject$getCciaAttr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="getCciaClass"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-getCciaClass'><code>cecelia::ReactivePersistentObject$getCciaClass()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="getCciaLogfile"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-getCciaLogfile'><code>cecelia::ReactivePersistentObject$getCciaLogfile()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="getCciaMeta"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-getCciaMeta'><code>cecelia::ReactivePersistentObject$getCciaMeta()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="getCciaName"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-getCciaName'><code>cecelia::ReactivePersistentObject$getCciaName()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="getCciaType"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-getCciaType'><code>cecelia::ReactivePersistentObject$getCciaType()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="getUID"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-getUID'><code>cecelia::ReactivePersistentObject$getUID()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="isTransaction"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-isTransaction'><code>cecelia::ReactivePersistentObject$isTransaction()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="lockReleased"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-lockReleased'><code>cecelia::ReactivePersistentObject$lockReleased()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="logfile"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-logfile'><code>cecelia::ReactivePersistentObject$logfile()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="logfileNames"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-logfileNames'><code>cecelia::ReactivePersistentObject$logfileNames()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="moduleFunParams"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-moduleFunParams'><code>cecelia::ReactivePersistentObject$moduleFunParams()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="objFiles"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-objFiles'><code>cecelia::ReactivePersistentObject$objFiles()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="persistentObjectDirectory"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-persistentObjectDirectory'><code>cecelia::ReactivePersistentObject$persistentObjectDirectory()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="persistentObjectDirectoryFile"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-persistentObjectDirectoryFile'><code>cecelia::ReactivePersistentObject$persistentObjectDirectoryFile()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="retrieveState"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-retrieveState'><code>cecelia::ReactivePersistentObject$retrieveState()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="runTask"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-runTask'><code>cecelia::ReactivePersistentObject$runTask()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="saveModuleFunParams"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-saveModuleFunParams'><code>cecelia::ReactivePersistentObject$saveModuleFunParams()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="saveState"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-saveState'><code>cecelia::ReactivePersistentObject$saveState()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="setCciaAttr"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-setCciaAttr'><code>cecelia::ReactivePersistentObject$setCciaAttr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="setCciaClass"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-setCciaClass'><code>cecelia::ReactivePersistentObject$setCciaClass()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="setCciaLogfile"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-setCciaLogfile'><code>cecelia::ReactivePersistentObject$setCciaLogfile()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="setCciaMeta"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-setCciaMeta'><code>cecelia::ReactivePersistentObject$setCciaMeta()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="setCciaName"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-setCciaName'><code>cecelia::ReactivePersistentObject$setCciaName()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="setCciaType"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-setCciaType'><code>cecelia::ReactivePersistentObject$setCciaType()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="setUID"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-setUID'><code>cecelia::ReactivePersistentObject$setUID()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="taskConf"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-taskConf'><code>cecelia::ReactivePersistentObject$taskConf()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="version"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-version'><code>cecelia::ReactivePersistentObject$version()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-new"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-new}{}}}
\subsection{Method \code{new()}}{
Complete population DT


init
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$new(stateFile, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{stateFile}}{character with file path to state file}

\item{\code{popUtils}}{PopulationUtils}

\item{\code{popUtils}}{PopulationUtils}

\item{\code{cols}}{list of character}

\item{\code{uniqueLabels}}{boolean to force unique labels}

\item{\code{includeX}}{boolean to include 'X' from adata}

\item{\code{includeObs}}{boolean to include 'obs' from adata}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-summary"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-summary}{}}}
\subsection{Method \code{summary()}}{
Summary of image
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$summary(fields = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fields}}{list of character for fields to include}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-editChannelName"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-editChannelName}{}}}
\subsection{Method \code{editChannelName()}}{
Edit channel names
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$editChannelName(objChannelNum, objChannelName, invalidate = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{objChannelNum}}{integer for channel number}

\item{\code{objChannelName}}{character for channel name}

\item{\code{invalidate}}{boolean to invalidate object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-omeXML"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-omeXML}{}}}
\subsection{Method \code{omeXML()}}{
OME xml
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$omeXML(reset = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{reset}}{boolean to reset values}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-omeXMLPath"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-omeXMLPath}{}}}
\subsection{Method \code{omeXMLPath()}}{
Modify xml path if needed for OME
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$omeXMLPath(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{character of path}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-omeXMLPixels"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-omeXMLPixels}{}}}
\subsection{Method \code{omeXMLPixels()}}{
pixel information
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$omeXMLPixels(reset = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{reset}}{boolean to reset value}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-omeXMLTimelapseInfo"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-omeXMLTimelapseInfo}{}}}
\subsection{Method \code{omeXMLTimelapseInfo()}}{
Timelapse information
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$omeXMLTimelapseInfo()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-omeXMLPixelRes"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-omeXMLPixelRes}{}}}
\subsection{Method \code{omeXMLPixelRes()}}{
Pixel resolution
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$omeXMLPixelRes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-resetMetaFromFile"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-resetMetaFromFile}{}}}
\subsection{Method \code{resetMetaFromFile()}}{
Reset meta information from file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$resetMetaFromFile()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-tracksInfo"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-tracksInfo}{}}}
\subsection{Method \code{tracksInfo()}}{
Tracks info
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$tracksInfo(
  trackStatsNames,
  parentPop,
  pop = NULL,
  addPops = NULL,
  forceReload = FALSE,
  flushCache = FALSE,
  replaceNA = TRUE,
  quantiles = c(0.95, 0.05),
  colsToNormalise = c(),
  normPercentile = 0.998
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{trackStatsNames}}{list of character to define which stats to return}

\item{\code{parentPop}}{character to define parent population}

\item{\code{pop}}{character to define population}

\item{\code{addPops}}{character to define populations to add}

\item{\code{forceReload}}{boolean to force reload of tracks}

\item{\code{flushCache}}{boolean to flush cache}

\item{\code{replaceNA}}{boolean to replace NA}

\item{\code{quantiles}}{list of N=2 to define quantiles}

\item{\code{colsToNormalise}}{list of character to define columns to normalise}

\item{\code{normPercentile}}{numeric to define normalisation percentile 0-1}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-tracksMeasures"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-tracksMeasures}{}}}
\subsection{Method \code{tracksMeasures()}}{
Tracks measures
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$tracksMeasures(measures, pops, forceReload = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{measures}}{list of character to define measures}

\item{\code{pops}}{list of character to define populations}

\item{\code{forceReload}}{boolean to force reload of values}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-tracks"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-tracks}{}}}
\subsection{Method \code{tracks()}}{
Tracks
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$tracks(pop, forceReload = FALSE, minTracklength = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pop}}{character to define population}

\item{\code{forceReload}}{boolean to force reload of tracks}

\item{\code{minTracklength}}{integer for minimum track length}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-popStats"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-popStats}{}}}
\subsection{Method \code{popStats()}}{
Population stats
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$popStats(popType)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{popType}}{character of population type}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-labelPropsCols"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-labelPropsCols}{}}}
\subsection{Method \code{labelPropsCols()}}{
property columns from labels
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$labelPropsCols(
  valueNames = NULL,
  dataTypes = c("vars", "obs"),
  popType = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{valueNames}}{list of character to define value names}

\item{\code{dataTypes}}{list of character to define data types from adata}

\item{\code{popType}}{character for population type}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-labelProps"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-labelProps}{}}}
\subsection{Method \code{labelProps()}}{
Label props
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$labelProps(valueName = NULL, forceReload = FALSE, readOnly = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{valueName}}{character for value name}

\item{\code{forceReload}}{boolean to force reload data}

\item{\code{readOnly}}{boolean to load dataset read-only - not sure that works}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-regionProps"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-regionProps}{}}}
\subsection{Method \code{regionProps()}}{
Region props
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$regionProps(valueName = NULL, forceReload = FALSE, readOnly = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{valueName}}{character for value name}

\item{\code{forceReload}}{boolean to force reload data}

\item{\code{readOnly}}{boolean to load dataset read-only - not sure that works}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-neighbourProps"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-neighbourProps}{}}}
\subsection{Method \code{neighbourProps()}}{
Neighbour props
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$neighbourProps(
  valueName = NULL,
  forceReload = FALSE,
  readOnly = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{valueName}}{character for value name}

\item{\code{forceReload}}{boolean to force reload data}

\item{\code{readOnly}}{boolean to load dataset read-only - not sure that works}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-popUtils"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-popUtils}{}}}
\subsection{Method \code{popUtils()}}{
Population object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$popUtils(popType, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{popType}}{character for population type}

\item{\code{...}}{passed to population utils object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-popPathsAll"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-popPathsAll}{}}}
\subsection{Method \code{popPathsAll()}}{
All population paths as list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$popPathsAll(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{passed to self$popPaths}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-popPaths"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-popPaths}{}}}
\subsection{Method \code{popPaths()}}{
Population paths
add populations from population map
if there were filtered
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$popPaths(
  popType,
  includeFiltered = FALSE,
  filteredOnly = FALSE,
  includeRoot = FALSE,
  tracksOnly = FALSE,
  cellsOnly = TRUE,
  parentPops = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{popType}}{character for population type}

\item{\code{includeFiltered}}{boolean to include filtered populations}

\item{\code{filteredOnly}}{boolean to only return filtered populations}

\item{\code{includeRoot}}{boolean to include root}

\item{\code{tracksOnly}}{boolean to return tracks only}

\item{\code{cellsOnly}}{boolean to return cells only}

\item{\code{parentPops}}{list of character of parent populations}

\item{\code{...}}{passed to self$popUtils}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-neighboursDT"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-neighboursDT}{}}}
\subsection{Method \code{neighboursDT()}}{
Neighbours DT
TODO could this be done easier?
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$neighboursDT(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{passed to self$neighbourProps}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-popDT"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-popDT}{}}}
\subsection{Method \code{popDT()}}{
Population datatable
apply filtering populations if needed
TODO this got a bit complicated - can you clean this up?
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$popDT(
  popType,
  pops = NULL,
  cols = NULL,
  dropNA = FALSE,
  dropPop = FALSE,
  includeFiltered = FALSE,
  forceReload = FALSE,
  uniqueLabels = TRUE,
  flushCache = FALSE,
  replaceNA = FALSE,
  completeDT = TRUE,
  filterMeasures = NULL,
  includeX = FALSE,
  includeObs = TRUE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{popType}}{character for population type}

\item{\code{pops}}{list of character for populations}

\item{\code{cols}}{list of character for columns to include}

\item{\code{dropNA}}{boolean to drop NA}

\item{\code{dropPop}}{boolean to drop population}

\item{\code{includeFiltered}}{boolean to include filtered population}

\item{\code{forceReload}}{boolean to force reload data}

\item{\code{uniqueLabels}}{boolean to force unique labels}

\item{\code{flushCache}}{boolean to flush cache}

\item{\code{replaceNA}}{boolean to replace NA}

\item{\code{completeDT}}{boolean to complete data.table with label props}

\item{\code{filterMeasures}}{list of character to include filter measures}

\item{\code{includeX}}{boolean to include 'X' from adata}

\item{\code{...}}{passed to self$popUtils}

\item{\code{variable}}{boolean to include 'obs' from adata}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-savePops"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-savePops}{}}}
\subsection{Method \code{savePops()}}{
Save population labels to disk
python can use that to filter
the adata files wihtout being
aware of the populations themselves
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$savePops(
  popType,
  pops = NULL,
  includeFiltered = FALSE,
  flushCache = TRUE,
  completeDT = TRUE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{popType}}{character for population type}

\item{\code{pops}}{list of character for populations}

\item{\code{includeFiltered}}{boolean to include filtered population}

\item{\code{flushCache}}{boolean to flush cache}

\item{\code{completeDT}}{boolean to complete data.table with label props}

\item{\code{...}}{passed to self$popUtils(popType = popType)$savePops}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-pp3"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-pp3}{}}}
\subsection{Method \code{pp3()}}{
pp3 object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$pp3(windowPops = NULL, pops = NULL, usePops = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{windowPops}}{list of character to define window}

\item{\code{pops}}{list of character to define populations}

\item{\code{usePops}}{list of character to filter population data.frame}

\item{\code{...}}{passed to self$popDT}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-ppp"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-ppp}{}}}
\subsection{Method \code{ppp()}}{
ppp
https://stackoverflow.com/a/39956181
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$ppp(windowPops = NULL, pops = NULL, usePops = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{windowPops}}{list of character to define window}

\item{\code{pops}}{list of character to define populations}

\item{\code{usePops}}{list of character to filter population data.frame}

\item{\code{...}}{passed to self$popDT}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-spatialDT"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-spatialDT}{}}}
\subsection{Method \code{spatialDT()}}{
Spatial DT
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$spatialDT(valueName = "default", forceReload = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{valueName}}{character for value name}

\item{\code{forceReload}}{boolean to force reload}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-spatialGraph"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-spatialGraph}{}}}
\subsection{Method \code{spatialGraph()}}{
Spatial igraph
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$spatialGraph(popType, valueName = "default", ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{popType}}{character for population type}

\item{\code{valueName}}{character for value name}

\item{\code{...}}{passed to self$spatialDT}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-flowFrame"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-flowFrame}{}}}
\subsection{Method \code{flowFrame()}}{
FlowFrame
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$flowFrame(
  valueName,
  compensateZ = FALSE,
  polyDegree = 4,
  applyReversedLog = FALSE,
  reversedLogBase = 0
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{valueName}}{character for value name}

\item{\code{compensateZ}}{boolean to compensate depth}

\item{\code{polyDegree}}{integer for plynomial compensation}

\item{\code{applyReversedLog}}{boolean to apply reverse log}

\item{\code{reversedLogBase}}{integer for base of reverse log}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-flowGatingSet"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-flowGatingSet}{}}}
\subsection{Method \code{flowGatingSet()}}{
Gating set
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$flowGatingSet(forceReload = FALSE, init = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{forceReload}}{boolean to force reload data}

\item{\code{init}}{boolean to init data}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-clsfPopUtils"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-clsfPopUtils}{}}}
\subsection{Method \code{clsfPopUtils()}}{
Classification utils
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$clsfPopUtils(forceReload = FALSE, init = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{forceReload}}{boolean to force reload data}

\item{\code{init}}{boolean to init data}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-livePopUtils"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-livePopUtils}{}}}
\subsection{Method \code{livePopUtils()}}{
Live population utils
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$livePopUtils(forceReload = FALSE, init = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{forceReload}}{boolean to force reload data}

\item{\code{init}}{boolean to init data}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-adataUtils"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-adataUtils}{}}}
\subsection{Method \code{adataUtils()}}{
Adata
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$adataUtils(forceReload = FALSE, init = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{forceReload}}{boolean to force reload data}

\item{\code{init}}{boolean to init data}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-savePopMap"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-savePopMap}{}}}
\subsection{Method \code{savePopMap()}}{
Save pop map to file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$savePopMap(popType, invalidate = TRUE, includeFiltered = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{popType}}{character for population type}

\item{\code{invalidate}}{boolean to invalidate object}

\item{\code{includeFiltered}}{boolean to include filtered populations}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-addFilteredPops"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-addFilteredPops}{}}}
\subsection{Method \code{addFilteredPops()}}{
Add filtered populations
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$addFilteredPops(
  popType,
  parentPops,
  pops,
  valueName = "default",
  tracksOnly = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{popType}}{character for population type}

\item{\code{parentPops}}{list of character for parent populations}

\item{\code{pops}}{list of character for populations}

\item{\code{valueName}}{character for value name}

\item{\code{tracksOnly}}{boolean to return tracks onlye}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-addPop"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-addPop}{}}}
\subsection{Method \code{addPop()}}{
Add population
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$addPop(
  popType,
  popName = NULL,
  popAttr = list(),
  invalidate = TRUE,
  includeFiltered = FALSE,
  popColour = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{popType}}{character for population type}

\item{\code{popName}}{character for population name}

\item{\code{popAttr}}{list of character for population attributes}

\item{\code{invalidate}}{boolean to invalidate object}

\item{\code{includeFiltered}}{boolean to include filtered populations}

\item{\code{popColour}}{character for population colour}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-setPopAttr"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-setPopAttr}{}}}
\subsection{Method \code{setPopAttr()}}{
Set attributes for population
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$setPopAttr(
  popType,
  popID,
  popAttr,
  invalidate = TRUE,
  includeFiltered = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{popType}}{character for population type}

\item{\code{popID}}{character for population ID}

\item{\code{popAttr}}{character for population attribute}

\item{\code{invalidate}}{boolean to invalidate object}

\item{\code{includeFiltered}}{boolean to include filtered populations}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-popAttr"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-popAttr}{}}}
\subsection{Method \code{popAttr()}}{
Pop attribute
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$popAttr(
  popType,
  popAttr,
  popIDs = NULL,
  popPath = NULL,
  includeFiltered = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{popType}}{character for population type}

\item{\code{popAttr}}{character for population attribute}

\item{\code{popIDs}}{list of character for population IDs}

\item{\code{popPath}}{character for population path}

\item{\code{includeFiltered}}{boolean to include filtered populations}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-popIDFromPath"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-popIDFromPath}{}}}
\subsection{Method \code{popIDFromPath()}}{
Pop id from path
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$popIDFromPath(popMap, popPath)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{popMap}}{list of pop map}

\item{\code{popPath}}{character for population path}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-popIDsFromPaths"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-popIDsFromPaths}{}}}
\subsection{Method \code{popIDsFromPaths()}}{
Pop ids from paths
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$popIDsFromPaths(popMap, popPaths)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{popMap}}{list of pop map}

\item{\code{popPaths}}{list of character for population paths}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-popLeaves"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-popLeaves}{}}}
\subsection{Method \code{popLeaves()}}{
Pop leaves
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$popLeaves(popType, pop, direct = FALSE, includeFiltered = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{popType}}{character for population type}

\item{\code{pop}}{character for population}

\item{\code{direct}}{boolean for direct population leaves}

\item{\code{includeFiltered}}{boolean to include filtered population}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-popIDsByAttrList"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-popIDsByAttrList}{}}}
\subsection{Method \code{popIDsByAttrList()}}{
Pop IDs by attribut list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$popIDsByAttrList(
  popType,
  popAttrList,
  combineLogic = "AND",
  compareFun = "eq",
  includeFiltered = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{popType}}{character for population type}

\item{\code{popAttrList}}{list of character for population attributes}

\item{\code{combineLogic}}{character for combine logic}

\item{\code{compareFun}}{character for compare function}

\item{\code{includeFiltered}}{boolean to include filtered populations}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-popIDsByAttr"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-popIDsByAttr}{}}}
\subsection{Method \code{popIDsByAttr()}}{
Pop IDs by attribute
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$popIDsByAttr(
  popType,
  popAttrKey,
  popAttrValue,
  compareFun = "eq",
  includeFiltered = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{popType}}{character for population type}

\item{\code{popAttrKey}}{character of attribute key}

\item{\code{popAttrValue}}{character of attribute value}

\item{\code{compareFun}}{character for compare function}

\item{\code{includeFiltered}}{boolean to include filtered populations}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-popIDByAttr"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-popIDByAttr}{}}}
\subsection{Method \code{popIDByAttr()}}{
Pop ID by attribute
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$popIDByAttr(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{passed to self$popIDsByAttr}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-delPop"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-delPop}{}}}
\subsection{Method \code{delPop()}}{
Delete population
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$delPop(
  popType,
  popIDs = NULL,
  popPath = NULL,
  invalidate = TRUE,
  includeFiltered = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{popType}}{character for population type}

\item{\code{popIDs}}{list of character for population IDs}

\item{\code{popPath}}{character for population path}

\item{\code{invalidate}}{boolean to invalidate object}

\item{\code{includeFiltered}}{boolean to include filtered populations}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-delPopsByPath"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-delPopsByPath}{}}}
\subsection{Method \code{delPopsByPath()}}{
Delete populations by path
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$delPopsByPath(
  popType,
  pops,
  invalidate = TRUE,
  includeFiltered = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{popType}}{character for population type}

\item{\code{pops}}{list of character for populations}

\item{\code{invalidate}}{boolean to invalidate object}

\item{\code{includeFiltered}}{boolean to include filtered populations}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-toggleVisibilityForPop"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-toggleVisibilityForPop}{}}}
\subsection{Method \code{toggleVisibilityForPop()}}{
Toggle visibility for pop
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$toggleVisibilityForPop(
  popType,
  popID,
  invalidate = TRUE,
  includeFiltered = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{popType}}{character for population type}

\item{\code{popID}}{character for population ID}

\item{\code{invalidate}}{boolean to invalidate object}

\item{\code{includeFiltered}}{boolean to include filtered populations}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-editPopName"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-editPopName}{}}}
\subsection{Method \code{editPopName()}}{
Rename population
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$editPopName(
  popType,
  popID,
  newPopName,
  invalidate = TRUE,
  includeFiltered = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{popType}}{character for population type}

\item{\code{popID}}{character for population ID}

\item{\code{newPopName}}{character for new population name}

\item{\code{invalidate}}{boolean to invalidate object}

\item{\code{includeFiltered}}{boolean to include filtered populations}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-editPopColour"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-editPopColour}{}}}
\subsection{Method \code{editPopColour()}}{
Set colour for population
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$editPopColour(
  popType,
  popID,
  colour,
  invalidate = TRUE,
  includeFiltered = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{popType}}{character for population type}

\item{\code{popID}}{character for population ID}

\item{\code{colour}}{character for population colour}

\item{\code{invalidate}}{boolean to invalidate object}

\item{\code{includeFiltered}}{boolean to include filtered populations}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-saveData"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-saveData}{}}}
\subsection{Method \code{saveData()}}{
Save data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$saveData()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-savePopsToLabels"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-savePopsToLabels}{}}}
\subsection{Method \code{savePopsToLabels()}}{
Save populations to labels
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$savePopsToLabels(popType)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{popType}}{character of population type}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-resetData"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-resetData}{}}}
\subsection{Method \code{resetData()}}{
Reset loaded data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$resetData()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-loadData"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-loadData}{}}}
\subsection{Method \code{loadData()}}{
Load data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$loadData(forceReload = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{forceReload}}{boolean to force reload data}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-values"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-values}{}}}
\subsection{Method \code{values()}}{
Values for versioned variable
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$values(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{character for variable to get from self$getCciaMeta()}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-valueDefault"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-valueDefault}{}}}
\subsection{Method \code{valueDefault()}}{
Value default for versioned variable
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$valueDefault(x, default = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{character for variable to get from self$getCciaMeta()}

\item{\code{default}}{boolean to use default value}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-valueNamesForPopType"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-valueNamesForPopType}{}}}
\subsection{Method \code{valueNamesForPopType()}}{
Value name for popType
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$valueNamesForPopType(popType, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{popType}}{character for population type}

\item{\code{...}}{passed to self$valueNames}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-valueNames"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-valueNames}{}}}
\subsection{Method \code{valueNames()}}{
Value names for versioned variable
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$valueNames(x, valueType = NULL, defaultOnly = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{character for variable to get from self$getCciaMeta()}

\item{\code{valueType}}{character for value type}

\item{\code{defaultOnly}}{boolean to return default only}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-valuePartOf"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-valuePartOf}{}}}
\subsection{Method \code{valuePartOf()}}{
Value part of for versioned variable
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$valuePartOf(x, valueName = "default")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{character of variable}

\item{\code{valueName}}{character of value name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-valueSuffixes"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-valueSuffixes}{}}}
\subsection{Method \code{valueSuffixes()}}{
Value suffixes for versioned variable
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$valueSuffixes(x, valueName = "default")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{character of variable}

\item{\code{valueName}}{character of value name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-valueLocUID"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-valueLocUID}{}}}
\subsection{Method \code{valueLocUID()}}{
Value locations for versioned variable
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$valueLocUID(x, valueName = "default")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{character of variable}

\item{\code{valueName}}{character of value name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-valueAttr"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-valueAttr}{}}}
\subsection{Method \code{valueAttr()}}{
Value attr for versioned variable
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$valueAttr(x, attrName, valueName = "default")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{character of variable}

\item{\code{valueName}}{character of value name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-setOriFilepath"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-setOriFilepath}{}}}
\subsection{Method \code{setOriFilepath()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$setOriFilepath(x, invalidate = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-setImFilepath"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-setImFilepath}{}}}
\subsection{Method \code{setImFilepath()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$setImFilepath(
  x,
  valueName = NULL,
  setDefault = TRUE,
  invalidate = TRUE,
  reset = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-setImMeta"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-setImMeta}{}}}
\subsection{Method \code{setImMeta()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$setImMeta(x, invalidate = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-setImChannelNames"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-setImChannelNames}{}}}
\subsection{Method \code{setImChannelNames()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$setImChannelNames(
  x,
  valueName = NULL,
  setDefault = TRUE,
  invalidate = TRUE,
  reset = FALSE,
  addNames = TRUE,
  checkLength = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-setImLabelsFilepath"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-setImLabelsFilepath}{}}}
\subsection{Method \code{setImLabelsFilepath()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$setImLabelsFilepath(
  x,
  valueName = NULL,
  setDefault = TRUE,
  invalidate = TRUE,
  reset = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-setImLabelPropsFilepath"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-setImLabelPropsFilepath}{}}}
\subsection{Method \code{setImLabelPropsFilepath()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$setImLabelPropsFilepath(
  x,
  valueName = NULL,
  setDefault = TRUE,
  invalidate = TRUE,
  reset = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-setImRegionsFilepath"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-setImRegionsFilepath}{}}}
\subsection{Method \code{setImRegionsFilepath()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$setImRegionsFilepath(
  x,
  valueName = NULL,
  setDefault = TRUE,
  invalidate = TRUE,
  reset = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-setImNeighboursFilepath"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-setImNeighboursFilepath}{}}}
\subsection{Method \code{setImNeighboursFilepath()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$setImNeighboursFilepath(
  x,
  valueName = NULL,
  setDefault = TRUE,
  invalidate = TRUE,
  reset = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-setImPopMap"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-setImPopMap}{}}}
\subsection{Method \code{setImPopMap()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$setImPopMap(popType, x, mergeMap = FALSE, invalidate = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-setImPopGatePlot"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-setImPopGatePlot}{}}}
\subsection{Method \code{setImPopGatePlot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$setImPopGatePlot(
  popType,
  plotID = NULL,
  plotParams = NULL,
  invalidate = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-setImGatingSetFilepath"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-setImGatingSetFilepath}{}}}
\subsection{Method \code{setImGatingSetFilepath()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$setImGatingSetFilepath(
  x,
  valueName = NULL,
  setDefault = TRUE,
  invalidate = TRUE,
  reset = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-setImAnndataFilepath"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-setImAnndataFilepath}{}}}
\subsection{Method \code{setImAnndataFilepath()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$setImAnndataFilepath(
  x,
  valueName = NULL,
  setDefault = TRUE,
  invalidate = TRUE,
  reset = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-imMeta"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-imMeta}{}}}
\subsection{Method \code{imMeta()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$imMeta()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-imChannelNames"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-imChannelNames}{}}}
\subsection{Method \code{imChannelNames()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$imChannelNames(
  valueName = NULL,
  useNames = TRUE,
  correctChannelNames = FALSE,
  includeTypes = FALSE,
  rmAttr = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-oriFilepath"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-oriFilepath}{}}}
\subsection{Method \code{oriFilepath()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$oriFilepath(modified = FALSE, revertToOri = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-imFilepath"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-imFilepath}{}}}
\subsection{Method \code{imFilepath()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$imFilepath(valueName = NULL, absolutePath = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-imLabelsFilepath"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-imLabelsFilepath}{}}}
\subsection{Method \code{imLabelsFilepath()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$imLabelsFilepath(valueName = NULL, absolutePath = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-imRegionsFilepath"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-imRegionsFilepath}{}}}
\subsection{Method \code{imRegionsFilepath()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$imRegionsFilepath(valueName = NULL, absolutePath = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-imNeighboursFilepath"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-imNeighboursFilepath}{}}}
\subsection{Method \code{imNeighboursFilepath()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$imNeighboursFilepath(valueName = NULL, absolutePath = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-imLabelPropsFilepath"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-imLabelPropsFilepath}{}}}
\subsection{Method \code{imLabelPropsFilepath()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$imLabelPropsFilepath(valueName = NULL, absolutePath = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-imGatingSetFilepath"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-imGatingSetFilepath}{}}}
\subsection{Method \code{imGatingSetFilepath()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$imGatingSetFilepath(valueName = NULL, absolutePath = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-imAnndataFilepath"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-imAnndataFilepath}{}}}
\subsection{Method \code{imAnndataFilepath()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$imAnndataFilepath(valueName = NULL, absolutePath = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-imShapes"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-imShapes}{}}}
\subsection{Method \code{imShapes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$imShapes(shapeType = "region", valueName = "default")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-imPopMap"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-imPopMap}{}}}
\subsection{Method \code{imPopMap()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$imPopMap(
  popType = NULL,
  popIDs = NULL,
  popPath = NULL,
  tracksOnly = FALSE,
  includeFiltered = FALSE,
  filterMeasures = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-imPopGatePlots"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-imPopGatePlots}{}}}
\subsection{Method \code{imPopGatePlots()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$imPopGatePlots(popType = NULL, plotID = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImage-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImage-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImage$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
