% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cciaImageCollection.R
\name{CciaImageCollection}
\alias{CciaImageCollection}
\title{Reactive image object collection}
\description{
Reactive image object collection
}
\examples{
TODO
}
\section{Super classes}{
\code{\link[cecelia:ReactiveObject]{cecelia::ReactiveObject}} -> \code{\link[cecelia:ReactivePersistentObject]{cecelia::ReactivePersistentObject}} -> \code{\link[cecelia:ReactivePersistentObjectSet]{cecelia::ReactivePersistentObjectSet}} -> \code{\link[cecelia:ReactivePersistentObjectCollection]{cecelia::ReactivePersistentObjectCollection}} -> \code{CciaImageCollection}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CciaImageCollection-new}{\code{CciaImageCollection$new()}}
\item \href{#method-CciaImageCollection-clFiles}{\code{CciaImageCollection$clFiles()}}
\item \href{#method-CciaImageCollection-modelFiles}{\code{CciaImageCollection$modelFiles()}}
\item \href{#method-CciaImageCollection-pixclFiles}{\code{CciaImageCollection$pixclFiles()}}
\item \href{#method-CciaImageCollection-n2vFiles}{\code{CciaImageCollection$n2vFiles()}}
\item \href{#method-CciaImageCollection-clone}{\code{CciaImageCollection$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactiveObject" data-id="reactive"><a href='../../cecelia/html/ReactiveObject.html#method-ReactiveObject-reactive'><code>cecelia::ReactiveObject$reactive()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactiveObject" data-id="update"><a href='../../cecelia/html/ReactiveObject.html#method-ReactiveObject-update'><code>cecelia::ReactiveObject$update()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="addCciaAttr"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-addCciaAttr'><code>cecelia::ReactivePersistentObject$addCciaAttr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="delCciaAttr"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-delCciaAttr'><code>cecelia::ReactivePersistentObject$delCciaAttr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="deleteObjectDirectory"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-deleteObjectDirectory'><code>cecelia::ReactivePersistentObject$deleteObjectDirectory()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="editCciaAttr"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-editCciaAttr'><code>cecelia::ReactivePersistentObject$editCciaAttr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="finalise"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-finalise'><code>cecelia::ReactivePersistentObject$finalise()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="getCciaAttr"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-getCciaAttr'><code>cecelia::ReactivePersistentObject$getCciaAttr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="getCciaClass"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-getCciaClass'><code>cecelia::ReactivePersistentObject$getCciaClass()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="getCciaLogfile"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-getCciaLogfile'><code>cecelia::ReactivePersistentObject$getCciaLogfile()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="getCciaMeta"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-getCciaMeta'><code>cecelia::ReactivePersistentObject$getCciaMeta()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="getCciaName"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-getCciaName'><code>cecelia::ReactivePersistentObject$getCciaName()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="getCciaType"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-getCciaType'><code>cecelia::ReactivePersistentObject$getCciaType()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="getUID"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-getUID'><code>cecelia::ReactivePersistentObject$getUID()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="isTransaction"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-isTransaction'><code>cecelia::ReactivePersistentObject$isTransaction()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="lockReleased"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-lockReleased'><code>cecelia::ReactivePersistentObject$lockReleased()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="logfile"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-logfile'><code>cecelia::ReactivePersistentObject$logfile()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="logfileNames"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-logfileNames'><code>cecelia::ReactivePersistentObject$logfileNames()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="moduleFunParams"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-moduleFunParams'><code>cecelia::ReactivePersistentObject$moduleFunParams()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="objFiles"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-objFiles'><code>cecelia::ReactivePersistentObject$objFiles()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="persistentObjectDirectory"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-persistentObjectDirectory'><code>cecelia::ReactivePersistentObject$persistentObjectDirectory()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="persistentObjectDirectoryFile"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-persistentObjectDirectoryFile'><code>cecelia::ReactivePersistentObject$persistentObjectDirectoryFile()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="runTask"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-runTask'><code>cecelia::ReactivePersistentObject$runTask()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="saveModuleFunParams"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-saveModuleFunParams'><code>cecelia::ReactivePersistentObject$saveModuleFunParams()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="setCciaAttr"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-setCciaAttr'><code>cecelia::ReactivePersistentObject$setCciaAttr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="setCciaClass"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-setCciaClass'><code>cecelia::ReactivePersistentObject$setCciaClass()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="setCciaLogfile"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-setCciaLogfile'><code>cecelia::ReactivePersistentObject$setCciaLogfile()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="setCciaMeta"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-setCciaMeta'><code>cecelia::ReactivePersistentObject$setCciaMeta()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="setCciaName"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-setCciaName'><code>cecelia::ReactivePersistentObject$setCciaName()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="setCciaType"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-setCciaType'><code>cecelia::ReactivePersistentObject$setCciaType()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="setUID"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-setUID'><code>cecelia::ReactivePersistentObject$setUID()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="taskConf"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-taskConf'><code>cecelia::ReactivePersistentObject$taskConf()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="version"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-version'><code>cecelia::ReactivePersistentObject$version()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObjectSet" data-id="addAttrForCciaObjects"><a href='../../cecelia/html/ReactivePersistentObjectSet.html#method-ReactivePersistentObjectSet-addAttrForCciaObjects'><code>cecelia::ReactivePersistentObjectSet$addAttrForCciaObjects()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObjectSet" data-id="addCciaObjects"><a href='../../cecelia/html/ReactivePersistentObjectSet.html#method-ReactivePersistentObjectSet-addCciaObjects'><code>cecelia::ReactivePersistentObjectSet$addCciaObjects()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObjectSet" data-id="cciaObjectByUID"><a href='../../cecelia/html/ReactivePersistentObjectSet.html#method-ReactivePersistentObjectSet-cciaObjectByUID'><code>cecelia::ReactivePersistentObjectSet$cciaObjectByUID()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObjectSet" data-id="cciaObjectMetas"><a href='../../cecelia/html/ReactivePersistentObjectSet.html#method-ReactivePersistentObjectSet-cciaObjectMetas'><code>cecelia::ReactivePersistentObjectSet$cciaObjectMetas()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObjectSet" data-id="cciaObjectNames"><a href='../../cecelia/html/ReactivePersistentObjectSet.html#method-ReactivePersistentObjectSet-cciaObjectNames'><code>cecelia::ReactivePersistentObjectSet$cciaObjectNames()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObjectSet" data-id="cciaObjectUIDs"><a href='../../cecelia/html/ReactivePersistentObjectSet.html#method-ReactivePersistentObjectSet-cciaObjectUIDs'><code>cecelia::ReactivePersistentObjectSet$cciaObjectUIDs()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObjectSet" data-id="cciaObjects"><a href='../../cecelia/html/ReactivePersistentObjectSet.html#method-ReactivePersistentObjectSet-cciaObjects'><code>cecelia::ReactivePersistentObjectSet$cciaObjects()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObjectSet" data-id="cciaObjectsAttrNames"><a href='../../cecelia/html/ReactivePersistentObjectSet.html#method-ReactivePersistentObjectSet-cciaObjectsAttrNames'><code>cecelia::ReactivePersistentObjectSet$cciaObjectsAttrNames()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObjectSet" data-id="cciaObjectsByAttr"><a href='../../cecelia/html/ReactivePersistentObjectSet.html#method-ReactivePersistentObjectSet-cciaObjectsByAttr'><code>cecelia::ReactivePersistentObjectSet$cciaObjectsByAttr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObjectSet" data-id="cciaObjectsByUIDs"><a href='../../cecelia/html/ReactivePersistentObjectSet.html#method-ReactivePersistentObjectSet-cciaObjectsByUIDs'><code>cecelia::ReactivePersistentObjectSet$cciaObjectsByUIDs()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObjectSet" data-id="delAttrForCciaObjects"><a href='../../cecelia/html/ReactivePersistentObjectSet.html#method-ReactivePersistentObjectSet-delAttrForCciaObjects'><code>cecelia::ReactivePersistentObjectSet$delAttrForCciaObjects()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObjectSet" data-id="editAttrForCciaObjects"><a href='../../cecelia/html/ReactivePersistentObjectSet.html#method-ReactivePersistentObjectSet-editAttrForCciaObjects'><code>cecelia::ReactivePersistentObjectSet$editAttrForCciaObjects()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObjectSet" data-id="getCciaObjectUIDs"><a href='../../cecelia/html/ReactivePersistentObjectSet.html#method-ReactivePersistentObjectSet-getCciaObjectUIDs'><code>cecelia::ReactivePersistentObjectSet$getCciaObjectUIDs()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObjectSet" data-id="removeCciaObjectByUID"><a href='../../cecelia/html/ReactivePersistentObjectSet.html#method-ReactivePersistentObjectSet-removeCciaObjectByUID'><code>cecelia::ReactivePersistentObjectSet$removeCciaObjectByUID()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObjectSet" data-id="removeCciaObjects"><a href='../../cecelia/html/ReactivePersistentObjectSet.html#method-ReactivePersistentObjectSet-removeCciaObjects'><code>cecelia::ReactivePersistentObjectSet$removeCciaObjects()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObjectSet" data-id="retrieveState"><a href='../../cecelia/html/ReactivePersistentObjectSet.html#method-ReactivePersistentObjectSet-retrieveState'><code>cecelia::ReactivePersistentObjectSet$retrieveState()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObjectSet" data-id="runTasks"><a href='../../cecelia/html/ReactivePersistentObjectSet.html#method-ReactivePersistentObjectSet-runTasks'><code>cecelia::ReactivePersistentObjectSet$runTasks()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObjectSet" data-id="saveState"><a href='../../cecelia/html/ReactivePersistentObjectSet.html#method-ReactivePersistentObjectSet-saveState'><code>cecelia::ReactivePersistentObjectSet$saveState()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObjectSet" data-id="setCciaObjectUIDs"><a href='../../cecelia/html/ReactivePersistentObjectSet.html#method-ReactivePersistentObjectSet-setCciaObjectUIDs'><code>cecelia::ReactivePersistentObjectSet$setCciaObjectUIDs()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObjectSet" data-id="summary"><a href='../../cecelia/html/ReactivePersistentObjectSet.html#method-ReactivePersistentObjectSet-summary'><code>cecelia::ReactivePersistentObjectSet$summary()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObjectCollection" data-id="addCciaObjectSet"><a href='../../cecelia/html/ReactivePersistentObjectCollection.html#method-ReactivePersistentObjectCollection-addCciaObjectSet'><code>cecelia::ReactivePersistentObjectCollection$addCciaObjectSet()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObjectCollection" data-id="addCciaObjectsToSet"><a href='../../cecelia/html/ReactivePersistentObjectCollection.html#method-ReactivePersistentObjectCollection-addCciaObjectsToSet'><code>cecelia::ReactivePersistentObjectCollection$addCciaObjectsToSet()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObjectCollection" data-id="cciaNodeObjectByUID"><a href='../../cecelia/html/ReactivePersistentObjectCollection.html#method-ReactivePersistentObjectCollection-cciaNodeObjectByUID'><code>cecelia::ReactivePersistentObjectCollection$cciaNodeObjectByUID()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObjectCollection" data-id="cciaObjectSetByName"><a href='../../cecelia/html/ReactivePersistentObjectCollection.html#method-ReactivePersistentObjectCollection-cciaObjectSetByName'><code>cecelia::ReactivePersistentObjectCollection$cciaObjectSetByName()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImageCollection-new"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImageCollection-new}{}}}
\subsection{Method \code{new()}}{
Init
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImageCollection$new(stateFile, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{stateFile}}{character with file path to state file}

\item{\code{...}}{passed to super$initialize}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImageCollection-clFiles"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImageCollection-clFiles}{}}}
\subsection{Method \code{clFiles()}}{
Classifier files
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImageCollection$clFiles(clType, fullPath = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{clType}}{character of classifier type}

\item{\code{fullPath}}{boolean to return full path}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImageCollection-modelFiles"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImageCollection-modelFiles}{}}}
\subsection{Method \code{modelFiles()}}{
Model files
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImageCollection$modelFiles(modelType, fullPath = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{modelType}}{character of model type}

\item{\code{fullPath}}{boolean to return full path}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImageCollection-pixclFiles"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImageCollection-pixclFiles}{}}}
\subsection{Method \code{pixclFiles()}}{
Pixel classifier files
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImageCollection$pixclFiles(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{passed to self$clFiles}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImageCollection-n2vFiles"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImageCollection-n2vFiles}{}}}
\subsection{Method \code{n2vFiles()}}{
Noise2Void model files
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImageCollection$n2vFiles(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{passed to self$modelFiles}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImageCollection-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImageCollection-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImageCollection$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
