% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cciaHelpers.R
\name{.normDTCol}
\alias{.normDTCol}
\title{#' @description Rename channel columns
#' @param channelList list of character for channels
#' @param channelNames list of character for channels names
#' @param channelPattern character of channel pattern
#' @param flowNames boolean to convert names to 'flow save names'
#' @examples
#' TODO
#' @export
renameChannelColumns <- function(channelList, channelNames, channelPattern = "mean_intensity",
flowNames = TRUE) {correct names as 'flow' names

if (flowNames == TRUE) {
channelNames <- .flowCorrectChannelNames(channelNames)
}}
\usage{
.normDTCol(DT, col, normCol = col, percentile = 0.98)
}
\arguments{
\item{DT}{data.frame of population}

\item{col}{character of column}

\item{normCol}{character of new column name}

\item{percentile}{numeric of normalisation percentile}
}
\description{
Normalise DT column
}
\section{rename all channels matching the pattern}{

}

\examples{
TODO
}
