% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowGatingSet.R
\name{FlowGatingSet}
\alias{FlowGatingSet}
\title{Populations represented in GatingSet}
\description{
Populations represented in GatingSet
}
\examples{
TODO
}
\section{Super classes}{
\code{\link[cecelia:ReactiveObject]{cecelia::ReactiveObject}} -> \code{\link[cecelia:PopulationUtils]{cecelia::PopulationUtils}} -> \code{FlowGatingSet}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FlowGatingSet-new}{\code{FlowGatingSet$new()}}
\item \href{#method-FlowGatingSet-popPaths}{\code{FlowGatingSet$popPaths()}}
\item \href{#method-FlowGatingSet-popLabels}{\code{FlowGatingSet$popLabels()}}
\item \href{#method-FlowGatingSet-popGateNum}{\code{FlowGatingSet$popGateNum()}}
\item \href{#method-FlowGatingSet-popGateIDs}{\code{FlowGatingSet$popGateIDs()}}
\item \href{#method-FlowGatingSet-popGateID}{\code{FlowGatingSet$popGateID()}}
\item \href{#method-FlowGatingSet-popParent}{\code{FlowGatingSet$popParent()}}
\item \href{#method-FlowGatingSet-flistFromGateCoords}{\code{FlowGatingSet$flistFromGateCoords()}}
\item \href{#method-FlowGatingSet-addPop}{\code{FlowGatingSet$addPop()}}
\item \href{#method-FlowGatingSet-setPop}{\code{FlowGatingSet$setPop()}}
\item \href{#method-FlowGatingSet-getPopGate}{\code{FlowGatingSet$getPopGate()}}
\item \href{#method-FlowGatingSet-getPopStats}{\code{FlowGatingSet$getPopStats()}}
\item \href{#method-FlowGatingSet-delPop}{\code{FlowGatingSet$delPop()}}
\item \href{#method-FlowGatingSet-renamePop}{\code{FlowGatingSet$renamePop()}}
\item \href{#method-FlowGatingSet-popIsRoot}{\code{FlowGatingSet$popIsRoot()}}
\item \href{#method-FlowGatingSet-popsInLeaves}{\code{FlowGatingSet$popsInLeaves()}}
\item \href{#method-FlowGatingSet-popLeaves}{\code{FlowGatingSet$popLeaves()}}
\item \href{#method-FlowGatingSet-popDirectLeaves}{\code{FlowGatingSet$popDirectLeaves()}}
\item \href{#method-FlowGatingSet-popChannels}{\code{FlowGatingSet$popChannels()}}
\item \href{#method-FlowGatingSet-popGatedOnChannels}{\code{FlowGatingSet$popGatedOnChannels()}}
\item \href{#method-FlowGatingSet-popGate}{\code{FlowGatingSet$popGate()}}
\item \href{#method-FlowGatingSet-defaultPops}{\code{FlowGatingSet$defaultPops()}}
\item \href{#method-FlowGatingSet-popDT}{\code{FlowGatingSet$popDT()}}
\item \href{#method-FlowGatingSet-changeParentName}{\code{FlowGatingSet$changeParentName()}}
\item \href{#method-FlowGatingSet-recompute}{\code{FlowGatingSet$recompute()}}
\item \href{#method-FlowGatingSet-copyGatesFrom}{\code{FlowGatingSet$copyGatesFrom()}}
\item \href{#method-FlowGatingSet-save}{\code{FlowGatingSet$save()}}
\item \href{#method-FlowGatingSet-isLabelPropsStore}{\code{FlowGatingSet$isLabelPropsStore()}}
\item \href{#method-FlowGatingSet-clone}{\code{FlowGatingSet$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactiveObject" data-id="reactive"><a href='../../cecelia/html/ReactiveObject.html#method-ReactiveObject-reactive'><code>cecelia::ReactiveObject$reactive()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactiveObject" data-id="update"><a href='../../cecelia/html/ReactiveObject.html#method-ReactiveObject-update'><code>cecelia::ReactiveObject$update()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="PopulationUtils" data-id="getImChannelLimits"><a href='../../cecelia/html/PopulationUtils.html#method-PopulationUtils-getImChannelLimits'><code>cecelia::PopulationUtils$getImChannelLimits()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="PopulationUtils" data-id="getImChannels"><a href='../../cecelia/html/PopulationUtils.html#method-PopulationUtils-getImChannels'><code>cecelia::PopulationUtils$getImChannels()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="PopulationUtils" data-id="getPopObj"><a href='../../cecelia/html/PopulationUtils.html#method-PopulationUtils-getPopObj'><code>cecelia::PopulationUtils$getPopObj()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="PopulationUtils" data-id="getPopObjFilepath"><a href='../../cecelia/html/PopulationUtils.html#method-PopulationUtils-getPopObjFilepath'><code>cecelia::PopulationUtils$getPopObjFilepath()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="PopulationUtils" data-id="getValueNames"><a href='../../cecelia/html/PopulationUtils.html#method-PopulationUtils-getValueNames'><code>cecelia::PopulationUtils$getValueNames()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="PopulationUtils" data-id="savePops"><a href='../../cecelia/html/PopulationUtils.html#method-PopulationUtils-savePops'><code>cecelia::PopulationUtils$savePops()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="PopulationUtils" data-id="unlinkPops"><a href='../../cecelia/html/PopulationUtils.html#method-PopulationUtils-unlinkPops'><code>cecelia::PopulationUtils$unlinkPops()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlowGatingSet-new"></a>}}
\if{latex}{\out{\hypertarget{method-FlowGatingSet-new}{}}}
\subsection{Method \code{new()}}{
Add rectangle gate


Add path gate


Init
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlowGatingSet$new(objFilepath, imChannels, valueName = "default")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{objFilepath}}{character for object file path}

\item{\code{imChannels}}{list of character for channel names}

\item{\code{valueName}}{character for value name}

\item{\code{gateCoords}}{list of (N,2) gate coordinates}

\item{\code{x}}{character of 'X'-coordinate}

\item{\code{y}}{character of 'Y'-coordinate}

\item{\code{...}}{.flowPolygonGate}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlowGatingSet-popPaths"></a>}}
\if{latex}{\out{\hypertarget{method-FlowGatingSet-popPaths}{}}}
\subsection{Method \code{popPaths()}}{
Gated populations paths
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlowGatingSet$popPaths(includeRoot = FALSE, popOrder = "tsort")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{includeRoot}}{boolean to include root}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlowGatingSet-popLabels"></a>}}
\if{latex}{\out{\hypertarget{method-FlowGatingSet-popLabels}{}}}
\subsection{Method \code{popLabels()}}{
Population labels
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlowGatingSet$popLabels()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlowGatingSet-popGateNum"></a>}}
\if{latex}{\out{\hypertarget{method-FlowGatingSet-popGateNum}{}}}
\subsection{Method \code{popGateNum()}}{
Number of gates for population
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlowGatingSet$popGateNum(parentPop, channels)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{parentPop}}{character for parent population}

\item{\code{channels}}{list of character (N,2) for channels}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlowGatingSet-popGateIDs"></a>}}
\if{latex}{\out{\hypertarget{method-FlowGatingSet-popGateIDs}{}}}
\subsection{Method \code{popGateIDs()}}{
Gate IDs for population
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlowGatingSet$popGateIDs(popPath)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{popPath}}{character for population path}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlowGatingSet-popGateID"></a>}}
\if{latex}{\out{\hypertarget{method-FlowGatingSet-popGateID}{}}}
\subsection{Method \code{popGateID()}}{
Gate ID for population
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlowGatingSet$popGateID(popPath)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{popPath}}{character for population path}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlowGatingSet-popParent"></a>}}
\if{latex}{\out{\hypertarget{method-FlowGatingSet-popParent}{}}}
\subsection{Method \code{popParent()}}{
Parent for population
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlowGatingSet$popParent(pop, root = "", normaliseRoot = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pop}}{character for population path}

\item{\code{root}}{character to identify root population}

\item{\code{normaliseRoot}}{boolean to normalise root}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlowGatingSet-flistFromGateCoords"></a>}}
\if{latex}{\out{\hypertarget{method-FlowGatingSet-flistFromGateCoords}{}}}
\subsection{Method \code{flistFromGateCoords()}}{
Filtered list from gate coords
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlowGatingSet$flistFromGateCoords(gateCoords, x, y)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{gateCoords}}{list of (N,2) gate coordinates}

\item{\code{x}}{character of 'X'-coordinate}

\item{\code{y}}{character of 'Y'-coordinate}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlowGatingSet-addPop"></a>}}
\if{latex}{\out{\hypertarget{method-FlowGatingSet-addPop}{}}}
\subsection{Method \code{addPop()}}{
Add population
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlowGatingSet$addPop(
  popName,
  gateCoords,
  x,
  y,
  parentPop = "root",
  flist = NULL,
  invalidate = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{popName}}{character for population name}

\item{\code{gateCoords}}{list of (N,2) gate coordinates}

\item{\code{x}}{character of 'X'-coordinate}

\item{\code{y}}{character of 'Y'-coordinate}

\item{\code{parentPop}}{character for parent population}

\item{\code{flist}}{flow list object}

\item{\code{invalidate}}{boolean to invalidate object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlowGatingSet-setPop"></a>}}
\if{latex}{\out{\hypertarget{method-FlowGatingSet-setPop}{}}}
\subsection{Method \code{setPop()}}{
Set population
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlowGatingSet$setPop(
  popPath,
  gateCoords,
  x,
  y,
  flist = NULL,
  invalidate = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{popPath}}{character for population path}

\item{\code{gateCoords}}{list of (N,2) gate coordinates}

\item{\code{x}}{character of 'X'-coordinate}

\item{\code{y}}{character of 'Y'-coordinate}

\item{\code{flist}}{flow list object}

\item{\code{invalidate}}{boolean to invalidate object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlowGatingSet-getPopGate"></a>}}
\if{latex}{\out{\hypertarget{method-FlowGatingSet-getPopGate}{}}}
\subsection{Method \code{getPopGate()}}{
Get population gate
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlowGatingSet$getPopGate(popPath)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{popPath}}{character for population path}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlowGatingSet-getPopStats"></a>}}
\if{latex}{\out{\hypertarget{method-FlowGatingSet-getPopStats}{}}}
\subsection{Method \code{getPopStats()}}{
Get population stats
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlowGatingSet$getPopStats(popPath, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{popPath}}{character for population path}

\item{\code{...}}{passed to .flowStatsForPop}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlowGatingSet-delPop"></a>}}
\if{latex}{\out{\hypertarget{method-FlowGatingSet-delPop}{}}}
\subsection{Method \code{delPop()}}{
Delete population
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlowGatingSet$delPop(popPath, invalidate = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{popPath}}{character for population path}

\item{\code{invalidate}}{boolean to invalidate object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlowGatingSet-renamePop"></a>}}
\if{latex}{\out{\hypertarget{method-FlowGatingSet-renamePop}{}}}
\subsection{Method \code{renamePop()}}{
Rename population
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlowGatingSet$renamePop(popPath, popName, invalidate = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{popPath}}{character for population path}

\item{\code{popName}}{character for new population name}

\item{\code{invalidate}}{boolean to invalidate object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlowGatingSet-popIsRoot"></a>}}
\if{latex}{\out{\hypertarget{method-FlowGatingSet-popIsRoot}{}}}
\subsection{Method \code{popIsRoot()}}{
Is population root?
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlowGatingSet$popIsRoot(pop)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pop}}{character for population path}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlowGatingSet-popsInLeaves"></a>}}
\if{latex}{\out{\hypertarget{method-FlowGatingSet-popsInLeaves}{}}}
\subsection{Method \code{popsInLeaves()}}{
Does the population exists in leaves?
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlowGatingSet$popsInLeaves(
  pops,
  parent = "root",
  directLeaves = FALSE,
  includeParent = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pops}}{list of character for populations}

\item{\code{parent}}{character for parent population}

\item{\code{directLeaves}}{boolean to only include direct leaves}

\item{\code{includeParent}}{boolean to include parent}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlowGatingSet-popLeaves"></a>}}
\if{latex}{\out{\hypertarget{method-FlowGatingSet-popLeaves}{}}}
\subsection{Method \code{popLeaves()}}{
Leaves of population
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlowGatingSet$popLeaves(popPath)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{popPath}}{character for population}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlowGatingSet-popDirectLeaves"></a>}}
\if{latex}{\out{\hypertarget{method-FlowGatingSet-popDirectLeaves}{}}}
\subsection{Method \code{popDirectLeaves()}}{
Direct leaves of population
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlowGatingSet$popDirectLeaves(popPath, channels = NULL, namesOnly = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{popPath}}{character for population}

\item{\code{channels}}{list of character (2) for channels}

\item{\code{namesOnly}}{boolean to return only population names not full path}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlowGatingSet-popChannels"></a>}}
\if{latex}{\out{\hypertarget{method-FlowGatingSet-popChannels}{}}}
\subsection{Method \code{popChannels()}}{
Channels of population
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlowGatingSet$popChannels(popPath)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{popPath}}{character for population}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlowGatingSet-popGatedOnChannels"></a>}}
\if{latex}{\out{\hypertarget{method-FlowGatingSet-popGatedOnChannels}{}}}
\subsection{Method \code{popGatedOnChannels()}}{
Check whether population is gated on the channels
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlowGatingSet$popGatedOnChannels(popPath, channels)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{popPath}}{character for population}

\item{\code{channels}}{list of character (2) for channels}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlowGatingSet-popGate"></a>}}
\if{latex}{\out{\hypertarget{method-FlowGatingSet-popGate}{}}}
\subsection{Method \code{popGate()}}{
Get gate for population
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlowGatingSet$popGate(popPath)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{popPath}}{character for population}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlowGatingSet-defaultPops"></a>}}
\if{latex}{\out{\hypertarget{method-FlowGatingSet-defaultPops}{}}}
\subsection{Method \code{defaultPops()}}{
Default pops
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlowGatingSet$defaultPops()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlowGatingSet-popDT"></a>}}
\if{latex}{\out{\hypertarget{method-FlowGatingSet-popDT}{}}}
\subsection{Method \code{popDT()}}{
Population data.table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlowGatingSet$popDT(
  pops = "root",
  cols = NULL,
  dropNA = FALSE,
  dropPop = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pops}}{list of character for populations}

\item{\code{cols}}{list of character for columns}

\item{\code{dropNA}}{boolean to drop NA}

\item{\code{dropPop}}{boolean to drop population}

\item{\code{copyDT}}{boolean to copy data.table}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlowGatingSet-changeParentName"></a>}}
\if{latex}{\out{\hypertarget{method-FlowGatingSet-changeParentName}{}}}
\subsection{Method \code{changeParentName()}}{
Change parent name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlowGatingSet$changeParentName(popPath, parentPath)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{popPath}}{character for population}

\item{\code{parentPath}}{character for parent population}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlowGatingSet-recompute"></a>}}
\if{latex}{\out{\hypertarget{method-FlowGatingSet-recompute}{}}}
\subsection{Method \code{recompute()}}{
Recompute gating
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlowGatingSet$recompute(invalidate = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{invalidate}}{boolean to invalidate object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlowGatingSet-copyGatesFrom"></a>}}
\if{latex}{\out{\hypertarget{method-FlowGatingSet-copyGatesFrom}{}}}
\subsection{Method \code{copyGatesFrom()}}{
Copy from other GatingSet
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlowGatingSet$copyGatesFrom(
  gsFrom,
  removeAll = TRUE,
  recompute = TRUE,
  invalidate = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{gsFrom}}{FlowGatingSet to copy from}

\item{\code{removeAll}}{boolean to remove all populations}

\item{\code{recompute}}{boolean to recompute after}

\item{\code{invalidate}}{boolean to invalidate object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlowGatingSet-save"></a>}}
\if{latex}{\out{\hypertarget{method-FlowGatingSet-save}{}}}
\subsection{Method \code{save()}}{
Save gating set
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlowGatingSet$save()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlowGatingSet-isLabelPropsStore"></a>}}
\if{latex}{\out{\hypertarget{method-FlowGatingSet-isLabelPropsStore}{}}}
\subsection{Method \code{isLabelPropsStore()}}{
Return whether utils are based on label props
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlowGatingSet$isLabelPropsStore()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlowGatingSet-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FlowGatingSet-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlowGatingSet$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
