% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jupyterKernelUtils.R
\name{JupyterKernelUtils}
\alias{JupyterKernelUtils}
\title{Jupyter Kernel utils}
\description{
Jupyter Kernel utils
}
\examples{
TODO
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-JupyterKernelUtils-new}{\code{JupyterKernelUtils$new()}}
\item \href{#method-JupyterKernelUtils-printConsoleConn}{\code{JupyterKernelUtils$printConsoleConn()}}
\item \href{#method-JupyterKernelUtils-quitKernel}{\code{JupyterKernelUtils$quitKernel()}}
\item \href{#method-JupyterKernelUtils-kernelClient}{\code{JupyterKernelUtils$kernelClient()}}
\item \href{#method-JupyterKernelUtils-execute}{\code{JupyterKernelUtils$execute()}}
\item \href{#method-JupyterKernelUtils-getKernelConnectionFile}{\code{JupyterKernelUtils$getKernelConnectionFile()}}
\item \href{#method-JupyterKernelUtils-clone}{\code{JupyterKernelUtils$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JupyterKernelUtils-new"></a>}}
\if{latex}{\out{\hypertarget{method-JupyterKernelUtils-new}{}}}
\subsection{Method \code{new()}}{
Init
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JupyterKernelUtils$new(connectionFile = NULL, useConnectionFile = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{connectionFile}}{character for connection file path}

\item{\code{condaEnv}}{character for conda environment}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JupyterKernelUtils-printConsoleConn"></a>}}
\if{latex}{\out{\hypertarget{method-JupyterKernelUtils-printConsoleConn}{}}}
\subsection{Method \code{printConsoleConn()}}{
print console connection info
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JupyterKernelUtils$printConsoleConn()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JupyterKernelUtils-quitKernel"></a>}}
\if{latex}{\out{\hypertarget{method-JupyterKernelUtils-quitKernel}{}}}
\subsection{Method \code{quitKernel()}}{
quit kernel
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JupyterKernelUtils$quitKernel()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JupyterKernelUtils-kernelClient"></a>}}
\if{latex}{\out{\hypertarget{method-JupyterKernelUtils-kernelClient}{}}}
\subsection{Method \code{kernelClient()}}{
return kernel client
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JupyterKernelUtils$kernelClient()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JupyterKernelUtils-execute"></a>}}
\if{latex}{\out{\hypertarget{method-JupyterKernelUtils-execute}{}}}
\subsection{Method \code{execute()}}{
execute command
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JupyterKernelUtils$execute(cmd, silent = TRUE, execInteractive = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JupyterKernelUtils-getKernelConnectionFile"></a>}}
\if{latex}{\out{\hypertarget{method-JupyterKernelUtils-getKernelConnectionFile}{}}}
\subsection{Method \code{getKernelConnectionFile()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JupyterKernelUtils$getKernelConnectionFile()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JupyterKernelUtils-clone"></a>}}
\if{latex}{\out{\hypertarget{method-JupyterKernelUtils-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JupyterKernelUtils$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
