% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/napariUtils.R
\name{NapariUtils}
\alias{NapariUtils}
\title{Napari utils}
\description{
Napari utils
}
\examples{
TODO
}
\section{Super class}{
\code{\link[cecelia:JupyterKernelUtils]{cecelia::JupyterKernelUtils}} -> \code{NapariUtils}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NapariUtils-initNapari}{\code{NapariUtils$initNapari()}}
\item \href{#method-NapariUtils-addPixclPane}{\code{NapariUtils$addPixclPane()}}
\item \href{#method-NapariUtils-addAnimationPane}{\code{NapariUtils$addAnimationPane()}}
\item \href{#method-NapariUtils-openViewer}{\code{NapariUtils$openViewer()}}
\item \href{#method-NapariUtils-closeViewer}{\code{NapariUtils$closeViewer()}}
\item \href{#method-NapariUtils-clearViewer}{\code{NapariUtils$clearViewer()}}
\item \href{#method-NapariUtils-clearViewerOutput}{\code{NapariUtils$clearViewerOutput()}}
\item \href{#method-NapariUtils-clearViewerInput}{\code{NapariUtils$clearViewerInput()}}
\item \href{#method-NapariUtils-openImage}{\code{NapariUtils$openImage()}}
\item \href{#method-NapariUtils-resetLabelsScale}{\code{NapariUtils$resetLabelsScale()}}
\item \href{#method-NapariUtils-saveLabels}{\code{NapariUtils$saveLabels()}}
\item \href{#method-NapariUtils-showLabelsAll}{\code{NapariUtils$showLabelsAll()}}
\item \href{#method-NapariUtils-showLabels}{\code{NapariUtils$showLabels()}}
\item \href{#method-NapariUtils-highlightTracks}{\code{NapariUtils$highlightTracks()}}
\item \href{#method-NapariUtils-highlightLabels}{\code{NapariUtils$highlightLabels()}}
\item \href{#method-NapariUtils-showChannelIntensity}{\code{NapariUtils$showChannelIntensity()}}
\item \href{#method-NapariUtils-showLayer}{\code{NapariUtils$showLayer()}}
\item \href{#method-NapariUtils-hideLayer}{\code{NapariUtils$hideLayer()}}
\item \href{#method-NapariUtils-setPopPointsSize}{\code{NapariUtils$setPopPointsSize()}}
\item \href{#method-NapariUtils-showPopMapping}{\code{NapariUtils$showPopMapping()}}
\item \href{#method-NapariUtils-showCellNeighbours}{\code{NapariUtils$showCellNeighbours()}}
\item \href{#method-NapariUtils-saveShapes}{\code{NapariUtils$saveShapes()}}
\item \href{#method-NapariUtils-showShapes}{\code{NapariUtils$showShapes()}}
\item \href{#method-NapariUtils-popLayerName}{\code{NapariUtils$popLayerName()}}
\item \href{#method-NapariUtils-showPreview}{\code{NapariUtils$showPreview()}}
\item \href{#method-NapariUtils-setTaskDir}{\code{NapariUtils$setTaskDir()}}
\item \href{#method-NapariUtils-clone}{\code{NapariUtils$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="JupyterKernelUtils" data-id="execute"><a href='../../cecelia/html/JupyterKernelUtils.html#method-JupyterKernelUtils-execute'><code>cecelia::JupyterKernelUtils$execute()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="JupyterKernelUtils" data-id="getKernelConnectionFile"><a href='../../cecelia/html/JupyterKernelUtils.html#method-JupyterKernelUtils-getKernelConnectionFile'><code>cecelia::JupyterKernelUtils$getKernelConnectionFile()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="JupyterKernelUtils" data-id="initialize"><a href='../../cecelia/html/JupyterKernelUtils.html#method-JupyterKernelUtils-initialize'><code>cecelia::JupyterKernelUtils$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="JupyterKernelUtils" data-id="kernelClient"><a href='../../cecelia/html/JupyterKernelUtils.html#method-JupyterKernelUtils-kernelClient'><code>cecelia::JupyterKernelUtils$kernelClient()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="JupyterKernelUtils" data-id="printConsoleConn"><a href='../../cecelia/html/JupyterKernelUtils.html#method-JupyterKernelUtils-printConsoleConn'><code>cecelia::JupyterKernelUtils$printConsoleConn()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="JupyterKernelUtils" data-id="quitKernel"><a href='../../cecelia/html/JupyterKernelUtils.html#method-JupyterKernelUtils-quitKernel'><code>cecelia::JupyterKernelUtils$quitKernel()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NapariUtils-initNapari"></a>}}
\if{latex}{\out{\hypertarget{method-NapariUtils-initNapari}{}}}
\subsection{Method \code{initNapari()}}{
Init napari
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NapariUtils$initNapari(
  viewerOutputFile,
  viewerInputFile,
  execInteractive = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{viewerOutputFile}}{character for output file path}

\item{\code{viewerInputFile}}{character for input file path}

\item{\code{execInteractive}}{boolean to execute interactive}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NapariUtils-addPixclPane"></a>}}
\if{latex}{\out{\hypertarget{method-NapariUtils-addPixclPane}{}}}
\subsection{Method \code{addPixclPane()}}{
Add pixel classifier
TODO this also required the napari extension
Should this be installed here if not present?
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NapariUtils$addPixclPane(execInteractive = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{execInteractive}}{boolean to execute interactive}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NapariUtils-addAnimationPane"></a>}}
\if{latex}{\out{\hypertarget{method-NapariUtils-addAnimationPane}{}}}
\subsection{Method \code{addAnimationPane()}}{
Add animation pane
TODO this also required the napari extension
Should this be installed here if not present?
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NapariUtils$addAnimationPane(execInteractive = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{execInteractive}}{boolean to execute interactive}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NapariUtils-openViewer"></a>}}
\if{latex}{\out{\hypertarget{method-NapariUtils-openViewer}{}}}
\subsection{Method \code{openViewer()}}{
Open viewer
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NapariUtils$openViewer(execInteractive = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{execInteractive}}{boolean to execute interactive}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NapariUtils-closeViewer"></a>}}
\if{latex}{\out{\hypertarget{method-NapariUtils-closeViewer}{}}}
\subsection{Method \code{closeViewer()}}{
Close viewer
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NapariUtils$closeViewer(execInteractive = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{execInteractive}}{boolean to execute interactive}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NapariUtils-clearViewer"></a>}}
\if{latex}{\out{\hypertarget{method-NapariUtils-clearViewer}{}}}
\subsection{Method \code{clearViewer()}}{
Clear viewer
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NapariUtils$clearViewer(execInteractive = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{execInteractive}}{boolean to execute interactive}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NapariUtils-clearViewerOutput"></a>}}
\if{latex}{\out{\hypertarget{method-NapariUtils-clearViewerOutput}{}}}
\subsection{Method \code{clearViewerOutput()}}{
Clear viewer output
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NapariUtils$clearViewerOutput()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NapariUtils-clearViewerInput"></a>}}
\if{latex}{\out{\hypertarget{method-NapariUtils-clearViewerInput}{}}}
\subsection{Method \code{clearViewerInput()}}{
Clear viewer input
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NapariUtils$clearViewerInput()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NapariUtils-openImage"></a>}}
\if{latex}{\out{\hypertarget{method-NapariUtils-openImage}{}}}
\subsection{Method \code{openImage()}}{
Open image
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NapariUtils$openImage(
  imPath,
  useChannelAxis = TRUE,
  imChannelNames = NULL,
  channelColormaps = NULL,
  napariModule = NULL,
  asDask = TRUE,
  show3D = FALSE,
  multiscales = NULL,
  execInteractive = TRUE,
  layersVisible = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{imPath}}{character for image path}

\item{\code{useChannelAxis}}{boolean to use channel axis}

\item{\code{imChannelNames}}{list of character for channel names}

\item{\code{channelColormaps}}{list of character for channel colormaps}

\item{\code{napariModule}}{character to add napari module}

\item{\code{asDask}}{boolean to load image as dask}

\item{\code{show3D}}{boolean to show image in 3D}

\item{\code{multiscales}}{integer for multiscales}

\item{\code{execInteractive}}{boolean to execute interactive}

\item{\code{layersVisible}}{boolean to make layers visible}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NapariUtils-resetLabelsScale"></a>}}
\if{latex}{\out{\hypertarget{method-NapariUtils-resetLabelsScale}{}}}
\subsection{Method \code{resetLabelsScale()}}{
Reset labels scale
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NapariUtils$resetLabelsScale(layersStartswith = NULL, execInteractive = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{layersStartswith}}{character to match layers}

\item{\code{execInteractive}}{boolean to execute interactive}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NapariUtils-saveLabels"></a>}}
\if{latex}{\out{\hypertarget{method-NapariUtils-saveLabels}{}}}
\subsection{Method \code{saveLabels()}}{
Save labels
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NapariUtils$saveLabels(
  pathToFile,
  layerName = "Labels",
  excludeNames = NULL,
  notifyModuleID = NULL,
  execInteractive = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pathToFile}}{character for labels path}

\item{\code{layerName}}{character for layer name}

\item{\code{excludeNames}}{boolean to exclude names}

\item{\code{notifyModuleID}}{character to notify a module once labels are saved}

\item{\code{execInteractive}}{boolean to execute interactive}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NapariUtils-showLabelsAll"></a>}}
\if{latex}{\out{\hypertarget{method-NapariUtils-showLabelsAll}{}}}
\subsection{Method \code{showLabelsAll()}}{
Show labels all
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NapariUtils$showLabelsAll(
  valueNames,
  showLabels = TRUE,
  showPoints = TRUE,
  showTracks = TRUE,
  showPops = TRUE,
  showNeighbours = TRUE,
  asNpArray = FALSE,
  execInteractive = TRUE,
  labelSuffixes = list()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{valueNames}}{list of character for value names}

\item{\code{showLabels}}{boolean to show labels}

\item{\code{showPoints}}{boolean to show points}

\item{\code{showTracks}}{boolean to show tracks}

\item{\code{showPops}}{boolean to show populations}

\item{\code{showNeighbours}}{boolean to show neighbours}

\item{\code{asNpArray}}{boolean to load labels as numpy array for editing}

\item{\code{execInteractive}}{boolean to execute interactive}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NapariUtils-showLabels"></a>}}
\if{latex}{\out{\hypertarget{method-NapariUtils-showLabels}{}}}
\subsection{Method \code{showLabels()}}{
Show labels
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NapariUtils$showLabels(
  showLabels = TRUE,
  showPoints = TRUE,
  showTracks = TRUE,
  asNpArray = FALSE,
  valueName = NULL,
  labelSuffixes = list(),
  execInteractive = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{showLabels}}{boolean to show labels}

\item{\code{showPoints}}{boolean to show points}

\item{\code{showTracks}}{boolean to show tracks}

\item{\code{asNpArray}}{boolean to load labels as numpy array for editing}

\item{\code{valueName}}{character for value name}

\item{\code{labelSuffixes}}{character for label suffixes}

\item{\code{execInteractive}}{boolean to execute interactive}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NapariUtils-highlightTracks"></a>}}
\if{latex}{\out{\hypertarget{method-NapariUtils-highlightTracks}{}}}
\subsection{Method \code{highlightTracks()}}{
Highlight tracks
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NapariUtils$highlightTracks(
  valueName,
  trackIDs,
  name = "*Tracks",
  execInteractive = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{valueName}}{character for value name}

\item{\code{trackIDs}}{list of integer for track IDs}

\item{\code{name}}{character for layer name}

\item{\code{execInteractive}}{boolean to execute interactive}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NapariUtils-highlightLabels"></a>}}
\if{latex}{\out{\hypertarget{method-NapariUtils-highlightLabels}{}}}
\subsection{Method \code{highlightLabels()}}{
Highlight labels
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NapariUtils$highlightLabels(valueName, labelIDs, execInteractive = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{valueName}}{character for value name}

\item{\code{labelIDs}}{list of integer for label IDs}

\item{\code{execInteractive}}{boolean to execute interactive}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NapariUtils-showChannelIntensity"></a>}}
\if{latex}{\out{\hypertarget{method-NapariUtils-showChannelIntensity}{}}}
\subsection{Method \code{showChannelIntensity()}}{
Show channel intensities
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NapariUtils$showChannelIntensity(channelID, execInteractive = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{channelID}}{integer for channel id}

\item{\code{execInteractive}}{boolean to execute interactive}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NapariUtils-showLayer"></a>}}
\if{latex}{\out{\hypertarget{method-NapariUtils-showLayer}{}}}
\subsection{Method \code{showLayer()}}{
Show layer
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NapariUtils$showLayer(layerName, execInteractive = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{layerName}}{character for layer naem}

\item{\code{execInteractive}}{boolean to execute interactive}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NapariUtils-hideLayer"></a>}}
\if{latex}{\out{\hypertarget{method-NapariUtils-hideLayer}{}}}
\subsection{Method \code{hideLayer()}}{
Hide layer
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NapariUtils$hideLayer(layerName, execInteractive = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{layerName}}{character for layer naem}

\item{\code{execInteractive}}{boolean to execute interactive}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NapariUtils-setPopPointsSize"></a>}}
\if{latex}{\out{\hypertarget{method-NapariUtils-setPopPointsSize}{}}}
\subsection{Method \code{setPopPointsSize()}}{
Set points size for populations
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NapariUtils$setPopPointsSize(popType, pointSize, execInteractive = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{popType}}{character for population type}

\item{\code{pointSize}}{integer for point size}

\item{\code{execInteractive}}{boolean to execute interactive}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NapariUtils-showPopMapping"></a>}}
\if{latex}{\out{\hypertarget{method-NapariUtils-showPopMapping}{}}}
\subsection{Method \code{showPopMapping()}}{
Show population mapping
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NapariUtils$showPopMapping(
  popType,
  valueName = NULL,
  removePrevious = TRUE,
  filteredFromValueName = FALSE,
  pointsSize = 6,
  execInteractive = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{popType}}{character for population type}

\item{\code{valueName}}{character for value name}

\item{\code{removePrevious}}{boolean to remove previous populations}

\item{\code{filteredFromValueName}}{boolean to filter from value name}

\item{\code{execInteractive}}{boolean to execute interactive}

\item{\code{pointSize}}{integer for point size}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NapariUtils-showCellNeighbours"></a>}}
\if{latex}{\out{\hypertarget{method-NapariUtils-showCellNeighbours}{}}}
\subsection{Method \code{showCellNeighbours()}}{
Show cell neighbours
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NapariUtils$showCellNeighbours(
  popType,
  valueName = "default",
  removePrevious = TRUE,
  execInteractive = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{popType}}{character for population type}

\item{\code{valueName}}{character for value name}

\item{\code{removePrevious}}{boolean to remove previous populations}

\item{\code{execInteractive}}{boolean to execute interactive}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NapariUtils-saveShapes"></a>}}
\if{latex}{\out{\hypertarget{method-NapariUtils-saveShapes}{}}}
\subsection{Method \code{saveShapes()}}{
Save shapes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NapariUtils$saveShapes(
  shapeType = "region",
  valueName = "default",
  execInteractive = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{shapeType}}{character for shape type}

\item{\code{valueName}}{character for value name}

\item{\code{execInteractive}}{boolean to execute interactive}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NapariUtils-showShapes"></a>}}
\if{latex}{\out{\hypertarget{method-NapariUtils-showShapes}{}}}
\subsection{Method \code{showShapes()}}{
Show shapes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NapariUtils$showShapes(
  shapeType = "region",
  valueName = "default",
  removePrevious = TRUE,
  execInteractive = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{shapeType}}{character for shape type}

\item{\code{valueName}}{character for value name}

\item{\code{removePrevious}}{boolean to remove previous}

\item{\code{execInteractive}}{boolean to execute interactive}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NapariUtils-popLayerName"></a>}}
\if{latex}{\out{\hypertarget{method-NapariUtils-popLayerName}{}}}
\subsection{Method \code{popLayerName()}}{
Get population layer name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NapariUtils$popLayerName(popType, popName)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{popType}}{character for population type}

\item{\code{popName}}{character for population name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NapariUtils-showPreview"></a>}}
\if{latex}{\out{\hypertarget{method-NapariUtils-showPreview}{}}}
\subsection{Method \code{showPreview()}}{
Show preview
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NapariUtils$showPreview(
  imageArray,
  channelNames = NULL,
  asPoints = FALSE,
  size = 5,
  asLabels = FALSE,
  multiscale = NULL,
  execInteractive = TRUE,
  useScale = TRUE,
  useChannelAxis = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{imageArray}}{character for image array - this is a python variable name}

\item{\code{channelNames}}{list of character for channel names}

\item{\code{asPoints}}{boolean to show as points}

\item{\code{size}}{integer for size}

\item{\code{asLabels}}{boolean to show as labels}

\item{\code{multiscale}}{integer for multiscales}

\item{\code{execInteractive}}{boolean to execute interactive}

\item{\code{useScale}}{boolean to use scale}

\item{\code{useChannelAxis}}{boolean to use channel axis}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NapariUtils-setTaskDir"></a>}}
\if{latex}{\out{\hypertarget{method-NapariUtils-setTaskDir}{}}}
\subsection{Method \code{setTaskDir()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NapariUtils$setTaskDir(x, execInteractive = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NapariUtils-clone"></a>}}
\if{latex}{\out{\hypertarget{method-NapariUtils-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NapariUtils$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
