% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populationUtils.R
\name{PopulationUtils}
\alias{PopulationUtils}
\title{GEneric population utils that have to be implemented for individual use cases.}
\description{
Generic population utils
}
\examples{
TODO
}
\section{Super class}{
\code{\link[cecelia:ReactiveObject]{cecelia::ReactiveObject}} -> \code{PopulationUtils}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PopulationUtils-new}{\code{PopulationUtils$new()}}
\item \href{#method-PopulationUtils-savePops}{\code{PopulationUtils$savePops()}}
\item \href{#method-PopulationUtils-unlinkPops}{\code{PopulationUtils$unlinkPops()}}
\item \href{#method-PopulationUtils-popDT}{\code{PopulationUtils$popDT()}}
\item \href{#method-PopulationUtils-save}{\code{PopulationUtils$save()}}
\item \href{#method-PopulationUtils-popPaths}{\code{PopulationUtils$popPaths()}}
\item \href{#method-PopulationUtils-popLabels}{\code{PopulationUtils$popLabels()}}
\item \href{#method-PopulationUtils-defaultPops}{\code{PopulationUtils$defaultPops()}}
\item \href{#method-PopulationUtils-isLabelPropsStore}{\code{PopulationUtils$isLabelPropsStore()}}
\item \href{#method-PopulationUtils-getPopObj}{\code{PopulationUtils$getPopObj()}}
\item \href{#method-PopulationUtils-getPopObjFilepath}{\code{PopulationUtils$getPopObjFilepath()}}
\item \href{#method-PopulationUtils-getImChannels}{\code{PopulationUtils$getImChannels()}}
\item \href{#method-PopulationUtils-getImChannelLimits}{\code{PopulationUtils$getImChannelLimits()}}
\item \href{#method-PopulationUtils-getValueNames}{\code{PopulationUtils$getValueNames()}}
\item \href{#method-PopulationUtils-clone}{\code{PopulationUtils$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactiveObject" data-id="reactive"><a href='../../cecelia/html/ReactiveObject.html#method-ReactiveObject-reactive'><code>cecelia::ReactiveObject$reactive()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactiveObject" data-id="update"><a href='../../cecelia/html/ReactiveObject.html#method-ReactiveObject-update'><code>cecelia::ReactiveObject$update()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PopulationUtils-new"></a>}}
\if{latex}{\out{\hypertarget{method-PopulationUtils-new}{}}}
\subsection{Method \code{new()}}{
Init channel limits


Init
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationUtils$new(popObjFilepath, imChannels)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{popObjFilepath}}{character for object filepath}

\item{\code{imChannels}}{list fo character for channel names}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PopulationUtils-savePops"></a>}}
\if{latex}{\out{\hypertarget{method-PopulationUtils-savePops}{}}}
\subsection{Method \code{savePops()}}{
save populations as csv files
this should be done everytime the population
is changing - otherwise napari will
not know what the populations are

TODO this is probably not the best way to do this and it only evolved
because populations were originally only defined in a GatingSet with
flowWorkspace. Now that populations are also defined in Anndata - I should
probably think about a more flexible approach.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationUtils$savePops(
  pops,
  popsDir,
  popDT = NULL,
  purge = FALSE,
  labels = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pops}}{list of character for populations}

\item{\code{popsDir}}{character for population directory}

\item{\code{popDT}}{data.table for populations}

\item{\code{purge}}{boolean to purge populations}

\item{\code{labels}}{list of integer for labels}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PopulationUtils-unlinkPops"></a>}}
\if{latex}{\out{\hypertarget{method-PopulationUtils-unlinkPops}{}}}
\subsection{Method \code{unlinkPops()}}{
Unlink population csv files
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationUtils$unlinkPops(pops, popsDir)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pops}}{list of character for populations}

\item{\code{popsDir}}{character to population directory}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PopulationUtils-popDT"></a>}}
\if{latex}{\out{\hypertarget{method-PopulationUtils-popDT}{}}}
\subsection{Method \code{popDT()}}{
Population data.table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationUtils$popDT(
  pops = NULL,
  cols = NULL,
  dropNA = FALSE,
  dropPop = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pops}}{list of character for populations}

\item{\code{cols}}{list of character for columns}

\item{\code{dropNA}}{boolean to drop NA}

\item{\code{dropPop}}{boolean to drop population}

\item{\code{copyDT}}{boolean to copy data.table}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PopulationUtils-save"></a>}}
\if{latex}{\out{\hypertarget{method-PopulationUtils-save}{}}}
\subsection{Method \code{save()}}{
Save populations back to disk
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationUtils$save()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PopulationUtils-popPaths"></a>}}
\if{latex}{\out{\hypertarget{method-PopulationUtils-popPaths}{}}}
\subsection{Method \code{popPaths()}}{
Populations paths
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationUtils$popPaths(includeRoot = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{includeRoot}}{boolean to include root}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PopulationUtils-popLabels"></a>}}
\if{latex}{\out{\hypertarget{method-PopulationUtils-popLabels}{}}}
\subsection{Method \code{popLabels()}}{
Pop labels
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationUtils$popLabels()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PopulationUtils-defaultPops"></a>}}
\if{latex}{\out{\hypertarget{method-PopulationUtils-defaultPops}{}}}
\subsection{Method \code{defaultPops()}}{
Default pops
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationUtils$defaultPops()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PopulationUtils-isLabelPropsStore"></a>}}
\if{latex}{\out{\hypertarget{method-PopulationUtils-isLabelPropsStore}{}}}
\subsection{Method \code{isLabelPropsStore()}}{
Whether utils are based on label props
for clusters, pull information from the main DT
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationUtils$isLabelPropsStore()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PopulationUtils-getPopObj"></a>}}
\if{latex}{\out{\hypertarget{method-PopulationUtils-getPopObj}{}}}
\subsection{Method \code{getPopObj()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationUtils$getPopObj()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PopulationUtils-getPopObjFilepath"></a>}}
\if{latex}{\out{\hypertarget{method-PopulationUtils-getPopObjFilepath}{}}}
\subsection{Method \code{getPopObjFilepath()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationUtils$getPopObjFilepath()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PopulationUtils-getImChannels"></a>}}
\if{latex}{\out{\hypertarget{method-PopulationUtils-getImChannels}{}}}
\subsection{Method \code{getImChannels()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationUtils$getImChannels()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PopulationUtils-getImChannelLimits"></a>}}
\if{latex}{\out{\hypertarget{method-PopulationUtils-getImChannelLimits}{}}}
\subsection{Method \code{getImChannelLimits()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationUtils$getImChannelLimits()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PopulationUtils-getValueNames"></a>}}
\if{latex}{\out{\hypertarget{method-PopulationUtils-getValueNames}{}}}
\subsection{Method \code{getValueNames()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationUtils$getValueNames()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PopulationUtils-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PopulationUtils-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationUtils$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
