% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactivePersistentObject.R
\name{ReactivePersistentObject}
\alias{ReactivePersistentObject}
\title{Reactive object that can be saved as RDS file}
\description{
Persistent version of reactive object
}
\examples{
TODO
}
\section{Super class}{
\code{\link[ceceliaCore:ReactiveObject]{ceceliaCore::ReactiveObject}} -> \code{ReactivePersistentObject}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ReactivePersistentObject-new}{\code{ReactivePersistentObject$new()}}
\item \href{#method-ReactivePersistentObject-finalize}{\code{ReactivePersistentObject$finalize()}}
\item \href{#method-ReactivePersistentObject-saveState}{\code{ReactivePersistentObject$saveState()}}
\item \href{#method-ReactivePersistentObject-retrieveState}{\code{ReactivePersistentObject$retrieveState()}}
\item \href{#method-ReactivePersistentObject-runTask}{\code{ReactivePersistentObject$runTask()}}
\item \href{#method-ReactivePersistentObject-taskConf}{\code{ReactivePersistentObject$taskConf()}}
\item \href{#method-ReactivePersistentObject-version}{\code{ReactivePersistentObject$version()}}
\item \href{#method-ReactivePersistentObject-summary}{\code{ReactivePersistentObject$summary()}}
\item \href{#method-ReactivePersistentObject-persistentObjectDirectory}{\code{ReactivePersistentObject$persistentObjectDirectory()}}
\item \href{#method-ReactivePersistentObject-persistentObjectDirectoryFile}{\code{ReactivePersistentObject$persistentObjectDirectoryFile()}}
\item \href{#method-ReactivePersistentObject-deleteObjectDirectory}{\code{ReactivePersistentObject$deleteObjectDirectory()}}
\item \href{#method-ReactivePersistentObject-addCciaAttr}{\code{ReactivePersistentObject$addCciaAttr()}}
\item \href{#method-ReactivePersistentObject-editCciaAttr}{\code{ReactivePersistentObject$editCciaAttr()}}
\item \href{#method-ReactivePersistentObject-delCciaAttr}{\code{ReactivePersistentObject$delCciaAttr()}}
\item \href{#method-ReactivePersistentObject-moduleFunParams}{\code{ReactivePersistentObject$moduleFunParams()}}
\item \href{#method-ReactivePersistentObject-saveModuleFunParams}{\code{ReactivePersistentObject$saveModuleFunParams()}}
\item \href{#method-ReactivePersistentObject-objFiles}{\code{ReactivePersistentObject$objFiles()}}
\item \href{#method-ReactivePersistentObject-setCciaName}{\code{ReactivePersistentObject$setCciaName()}}
\item \href{#method-ReactivePersistentObject-setCciaType}{\code{ReactivePersistentObject$setCciaType()}}
\item \href{#method-ReactivePersistentObject-setCciaClass}{\code{ReactivePersistentObject$setCciaClass()}}
\item \href{#method-ReactivePersistentObject-setCciaMeta}{\code{ReactivePersistentObject$setCciaMeta()}}
\item \href{#method-ReactivePersistentObject-setCciaAttr}{\code{ReactivePersistentObject$setCciaAttr()}}
\item \href{#method-ReactivePersistentObject-setCciaLogfile}{\code{ReactivePersistentObject$setCciaLogfile()}}
\item \href{#method-ReactivePersistentObject-setUID}{\code{ReactivePersistentObject$setUID()}}
\item \href{#method-ReactivePersistentObject-getCciaName}{\code{ReactivePersistentObject$getCciaName()}}
\item \href{#method-ReactivePersistentObject-getCciaType}{\code{ReactivePersistentObject$getCciaType()}}
\item \href{#method-ReactivePersistentObject-getCciaClass}{\code{ReactivePersistentObject$getCciaClass()}}
\item \href{#method-ReactivePersistentObject-getCciaMeta}{\code{ReactivePersistentObject$getCciaMeta()}}
\item \href{#method-ReactivePersistentObject-getCciaAttr}{\code{ReactivePersistentObject$getCciaAttr()}}
\item \href{#method-ReactivePersistentObject-logfile}{\code{ReactivePersistentObject$logfile()}}
\item \href{#method-ReactivePersistentObject-logfileNames}{\code{ReactivePersistentObject$logfileNames()}}
\item \href{#method-ReactivePersistentObject-getCciaLogfile}{\code{ReactivePersistentObject$getCciaLogfile()}}
\item \href{#method-ReactivePersistentObject-getUID}{\code{ReactivePersistentObject$getUID()}}
\item \href{#method-ReactivePersistentObject-isTransaction}{\code{ReactivePersistentObject$isTransaction()}}
\item \href{#method-ReactivePersistentObject-lockReleased}{\code{ReactivePersistentObject$lockReleased()}}
\item \href{#method-ReactivePersistentObject-clone}{\code{ReactivePersistentObject$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="ReactiveObject" data-id="reactive"><a href='../../ceceliaCore/html/ReactiveObject.html#method-ReactiveObject-reactive'><code>ceceliaCore::ReactiveObject$reactive()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="ReactiveObject" data-id="update"><a href='../../ceceliaCore/html/ReactiveObject.html#method-ReactiveObject-update'><code>ceceliaCore::ReactiveObject$update()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactivePersistentObject-new"></a>}}
\if{latex}{\out{\hypertarget{method-ReactivePersistentObject-new}{}}}
\subsection{Method \code{new()}}{
return lockfile for naive transactions


start naive transaction


finish naive transaction


Set state file


Set whether to init transation


Set whether lock is released


Get state file


Initialise parameters with list


Function mapping name for parameters


Init object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactivePersistentObject$new(
  stateFile,
  uID = NULL,
  initParams = NULL,
  initReactivity = TRUE,
  initTransaction = FALSE,
  retrieveState = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{stateFile}}{File path to state file}

\item{\code{uID}}{Unique ID of object}

\item{\code{initParams}}{list}

\item{\code{initParams}}{list to initialise parameters}

\item{\code{initReactivity}}{boolean to init reactivity}

\item{\code{initTransaction}}{boolean to init transaction}

\item{\code{retrieveState}}{boolean to retrieve state from file}

\item{\code{x}}{character}

\item{\code{invalidate}}{boolean}

\item{\code{x}}{boolean}

\item{\code{invalidate}}{boolean}

\item{\code{x}}{boolean}

\item{\code{invalidate}}{boolean}

\item{\code{moduleFun}}{character}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactivePersistentObject-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-ReactivePersistentObject-finalize}{}}}
\subsection{Method \code{finalize()}}{
on finalise
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactivePersistentObject$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactivePersistentObject-saveState"></a>}}
\if{latex}{\out{\hypertarget{method-ReactivePersistentObject-saveState}{}}}
\subsection{Method \code{saveState()}}{
Save state
TODO this is the most time consuming bit of the whole framework
You should think about how to store data in a database
rather than individual objects?
Saving without compression does not seem to make a difference?
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactivePersistentObject$saveState(saveData = TRUE, compressRDS = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{saveData}}{boolean to save data}

\item{\code{compressRDS}}{boolean to compress RDS file}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactivePersistentObject-retrieveState"></a>}}
\if{latex}{\out{\hypertarget{method-ReactivePersistentObject-retrieveState}{}}}
\subsection{Method \code{retrieveState()}}{
Retrieve state
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactivePersistentObject$retrieveState(
  initTransaction = FALSE,
  invalidate = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{initTransaction}}{boolean to init transaction}

\item{\code{invalidate}}{boolean to invalidate object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactivePersistentObject-runTask"></a>}}
\if{latex}{\out{\hypertarget{method-ReactivePersistentObject-runTask}{}}}
\subsection{Method \code{runTask()}}{
Run task
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactivePersistentObject$runTask(
  funName,
  funParams = list(),
  taskID = NULL,
  uploadSubmissionFiles = TRUE,
  taskConf = NULL,
  runInplace = TRUE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{funName}}{character for function name}

\item{\code{funParams}}{list for functions parameters}

\item{\code{taskID}}{integer for task ID}

\item{\code{uploadSubmissionFiles}}{boolean to upload submission file to HPC}

\item{\code{taskConf}}{list for function configuration}

\item{\code{runInplace}}{boolean to run function in place}

\item{\code{...}}{self$taskConf}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactivePersistentObject-taskConf"></a>}}
\if{latex}{\out{\hypertarget{method-ReactivePersistentObject-taskConf}{}}}
\subsection{Method \code{taskConf()}}{
Task variables for object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactivePersistentObject$taskConf(
  funParams,
  envVars = list(),
  hpcDir = NULL,
  pID = "000000",
  pName = "NONE",
  env = "local"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{funParams}}{list for function params}

\item{\code{envVars}}{list for running environment params}

\item{\code{hpcDir}}{character for HPC directory}

\item{\code{pID}}{character for project ID}

\item{\code{pName}}{character for project name}

\item{\code{env}}{character for environment name. One of c("local", "hpc").}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactivePersistentObject-version"></a>}}
\if{latex}{\out{\hypertarget{method-ReactivePersistentObject-version}{}}}
\subsection{Method \code{version()}}{
Current version
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactivePersistentObject$version()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactivePersistentObject-summary"></a>}}
\if{latex}{\out{\hypertarget{method-ReactivePersistentObject-summary}{}}}
\subsection{Method \code{summary()}}{
Summary of object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactivePersistentObject$summary(fields = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactivePersistentObject-persistentObjectDirectory"></a>}}
\if{latex}{\out{\hypertarget{method-ReactivePersistentObject-persistentObjectDirectory}{}}}
\subsection{Method \code{persistentObjectDirectory()}}{
Path to persistent object directory
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactivePersistentObject$persistentObjectDirectory(
  root = FALSE,
  zero = FALSE,
  uID = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{root}}{boolean to get root directory}

\item{\code{zero}}{boolean to get zero directory}

\item{\code{uID}}{character for unique ID}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactivePersistentObject-persistentObjectDirectoryFile"></a>}}
\if{latex}{\out{\hypertarget{method-ReactivePersistentObject-persistentObjectDirectoryFile}{}}}
\subsection{Method \code{persistentObjectDirectoryFile()}}{
Files in object directory
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactivePersistentObject$persistentObjectDirectoryFile(filename, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filename}}{character for file name}

\item{\code{...}}{self$persistentObjectDirectory}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactivePersistentObject-deleteObjectDirectory"></a>}}
\if{latex}{\out{\hypertarget{method-ReactivePersistentObject-deleteObjectDirectory}{}}}
\subsection{Method \code{deleteObjectDirectory()}}{
Remove object directory
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactivePersistentObject$deleteObjectDirectory(removeZero = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{removeZero}}{boolean to remove zero}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactivePersistentObject-addCciaAttr"></a>}}
\if{latex}{\out{\hypertarget{method-ReactivePersistentObject-addCciaAttr}{}}}
\subsection{Method \code{addCciaAttr()}}{
Add attribute
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactivePersistentObject$addCciaAttr(
  objAttrName,
  objAttrVal,
  invalidate = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{objAttrName}}{character for attribute name}

\item{\code{objAttrVal}}{character for attribute value}

\item{\code{invalidate}}{boolean to invalidate object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactivePersistentObject-editCciaAttr"></a>}}
\if{latex}{\out{\hypertarget{method-ReactivePersistentObject-editCciaAttr}{}}}
\subsection{Method \code{editCciaAttr()}}{
Edit attribute
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactivePersistentObject$editCciaAttr(
  objAttrName,
  objAttrVal,
  addAttr = TRUE,
  invalidate = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{objAttrName}}{character for attribute name}

\item{\code{objAttrVal}}{character for attribute value}

\item{\code{addAttr}}{boolean to add attribute if not exists}

\item{\code{invalidate}}{boolean to invalidate object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactivePersistentObject-delCciaAttr"></a>}}
\if{latex}{\out{\hypertarget{method-ReactivePersistentObject-delCciaAttr}{}}}
\subsection{Method \code{delCciaAttr()}}{
Delete/Remove attribute
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactivePersistentObject$delCciaAttr(objAttrName, invalidate = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{objAttrName}}{character for attribute name}

\item{\code{invalidate}}{boolean to invalidate object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactivePersistentObject-moduleFunParams"></a>}}
\if{latex}{\out{\hypertarget{method-ReactivePersistentObject-moduleFunParams}{}}}
\subsection{Method \code{moduleFunParams()}}{
Function parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactivePersistentObject$moduleFunParams(moduleFun)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{moduleFun}}{character for module function}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactivePersistentObject-saveModuleFunParams"></a>}}
\if{latex}{\out{\hypertarget{method-ReactivePersistentObject-saveModuleFunParams}{}}}
\subsection{Method \code{saveModuleFunParams()}}{
Save parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactivePersistentObject$saveModuleFunParams(moduleFun, x, invalidate = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{moduleFun}}{character for module function}

\item{\code{x}}{list of params}

\item{\code{invalidate}}{boolean to invalidate object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactivePersistentObject-objFiles"></a>}}
\if{latex}{\out{\hypertarget{method-ReactivePersistentObject-objFiles}{}}}
\subsection{Method \code{objFiles()}}{
Object files
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactivePersistentObject$objFiles()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactivePersistentObject-setCciaName"></a>}}
\if{latex}{\out{\hypertarget{method-ReactivePersistentObject-setCciaName}{}}}
\subsection{Method \code{setCciaName()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactivePersistentObject$setCciaName(x, invalidate = TRUE, reset = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactivePersistentObject-setCciaType"></a>}}
\if{latex}{\out{\hypertarget{method-ReactivePersistentObject-setCciaType}{}}}
\subsection{Method \code{setCciaType()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactivePersistentObject$setCciaType(x, invalidate = TRUE, reset = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactivePersistentObject-setCciaClass"></a>}}
\if{latex}{\out{\hypertarget{method-ReactivePersistentObject-setCciaClass}{}}}
\subsection{Method \code{setCciaClass()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactivePersistentObject$setCciaClass(x, invalidate = TRUE, reset = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactivePersistentObject-setCciaMeta"></a>}}
\if{latex}{\out{\hypertarget{method-ReactivePersistentObject-setCciaMeta}{}}}
\subsection{Method \code{setCciaMeta()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactivePersistentObject$setCciaMeta(x, invalidate = TRUE, reset = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactivePersistentObject-setCciaAttr"></a>}}
\if{latex}{\out{\hypertarget{method-ReactivePersistentObject-setCciaAttr}{}}}
\subsection{Method \code{setCciaAttr()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactivePersistentObject$setCciaAttr(x, invalidate = TRUE, reset = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactivePersistentObject-setCciaLogfile"></a>}}
\if{latex}{\out{\hypertarget{method-ReactivePersistentObject-setCciaLogfile}{}}}
\subsection{Method \code{setCciaLogfile()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactivePersistentObject$setCciaLogfile(
  x,
  valueName = "default",
  setDefault = TRUE,
  invalidate = TRUE,
  reset = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactivePersistentObject-setUID"></a>}}
\if{latex}{\out{\hypertarget{method-ReactivePersistentObject-setUID}{}}}
\subsection{Method \code{setUID()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactivePersistentObject$setUID(x, invalidate = TRUE, reset = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactivePersistentObject-getCciaName"></a>}}
\if{latex}{\out{\hypertarget{method-ReactivePersistentObject-getCciaName}{}}}
\subsection{Method \code{getCciaName()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactivePersistentObject$getCciaName()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactivePersistentObject-getCciaType"></a>}}
\if{latex}{\out{\hypertarget{method-ReactivePersistentObject-getCciaType}{}}}
\subsection{Method \code{getCciaType()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactivePersistentObject$getCciaType()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactivePersistentObject-getCciaClass"></a>}}
\if{latex}{\out{\hypertarget{method-ReactivePersistentObject-getCciaClass}{}}}
\subsection{Method \code{getCciaClass()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactivePersistentObject$getCciaClass()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactivePersistentObject-getCciaMeta"></a>}}
\if{latex}{\out{\hypertarget{method-ReactivePersistentObject-getCciaMeta}{}}}
\subsection{Method \code{getCciaMeta()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactivePersistentObject$getCciaMeta()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactivePersistentObject-getCciaAttr"></a>}}
\if{latex}{\out{\hypertarget{method-ReactivePersistentObject-getCciaAttr}{}}}
\subsection{Method \code{getCciaAttr()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactivePersistentObject$getCciaAttr(attrName = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactivePersistentObject-logfile"></a>}}
\if{latex}{\out{\hypertarget{method-ReactivePersistentObject-logfile}{}}}
\subsection{Method \code{logfile()}}{
Current or requested logfile only
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactivePersistentObject$logfile(valueName = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{valueName}}{character for value name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactivePersistentObject-logfileNames"></a>}}
\if{latex}{\out{\hypertarget{method-ReactivePersistentObject-logfileNames}{}}}
\subsection{Method \code{logfileNames()}}{
Current all logfile names
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactivePersistentObject$logfileNames()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactivePersistentObject-getCciaLogfile"></a>}}
\if{latex}{\out{\hypertarget{method-ReactivePersistentObject-getCciaLogfile}{}}}
\subsection{Method \code{getCciaLogfile()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactivePersistentObject$getCciaLogfile()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactivePersistentObject-getUID"></a>}}
\if{latex}{\out{\hypertarget{method-ReactivePersistentObject-getUID}{}}}
\subsection{Method \code{getUID()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactivePersistentObject$getUID()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactivePersistentObject-isTransaction"></a>}}
\if{latex}{\out{\hypertarget{method-ReactivePersistentObject-isTransaction}{}}}
\subsection{Method \code{isTransaction()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactivePersistentObject$isTransaction()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactivePersistentObject-lockReleased"></a>}}
\if{latex}{\out{\hypertarget{method-ReactivePersistentObject-lockReleased}{}}}
\subsection{Method \code{lockReleased()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactivePersistentObject$lockReleased()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactivePersistentObject-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ReactivePersistentObject-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactivePersistentObject$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
