% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactivePersistentObjectCollection.R
\name{ReactivePersistentObjectCollection}
\alias{ReactivePersistentObjectCollection}
\title{Collection of reactive persistent object sets.}
\description{
Collection of persistent reactive object sets
}
\examples{
TODO
}
\section{Super classes}{
\code{\link[cecelia:ReactiveObject]{cecelia::ReactiveObject}} -> \code{\link[cecelia:ReactivePersistentObject]{cecelia::ReactivePersistentObject}} -> \code{\link[cecelia:ReactivePersistentObjectSet]{cecelia::ReactivePersistentObjectSet}} -> \code{ReactivePersistentObjectCollection}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ReactivePersistentObjectCollection-addCciaObjectsToSet}{\code{ReactivePersistentObjectCollection$addCciaObjectsToSet()}}
\item \href{#method-ReactivePersistentObjectCollection-addCciaObjectSet}{\code{ReactivePersistentObjectCollection$addCciaObjectSet()}}
\item \href{#method-ReactivePersistentObjectCollection-cciaObjectSetByName}{\code{ReactivePersistentObjectCollection$cciaObjectSetByName()}}
\item \href{#method-ReactivePersistentObjectCollection-cciaNodeObjectByUID}{\code{ReactivePersistentObjectCollection$cciaNodeObjectByUID()}}
\item \href{#method-ReactivePersistentObjectCollection-clone}{\code{ReactivePersistentObjectCollection$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactiveObject" data-id="reactive"><a href='../../cecelia/html/ReactiveObject.html#method-ReactiveObject-reactive'><code>cecelia::ReactiveObject$reactive()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactiveObject" data-id="update"><a href='../../cecelia/html/ReactiveObject.html#method-ReactiveObject-update'><code>cecelia::ReactiveObject$update()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="addCciaAttr"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-addCciaAttr'><code>cecelia::ReactivePersistentObject$addCciaAttr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="delCciaAttr"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-delCciaAttr'><code>cecelia::ReactivePersistentObject$delCciaAttr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="deleteObjectDirectory"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-deleteObjectDirectory'><code>cecelia::ReactivePersistentObject$deleteObjectDirectory()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="editCciaAttr"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-editCciaAttr'><code>cecelia::ReactivePersistentObject$editCciaAttr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="finalize"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-finalize'><code>cecelia::ReactivePersistentObject$finalize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="getCciaAttr"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-getCciaAttr'><code>cecelia::ReactivePersistentObject$getCciaAttr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="getCciaClass"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-getCciaClass'><code>cecelia::ReactivePersistentObject$getCciaClass()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="getCciaLogfile"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-getCciaLogfile'><code>cecelia::ReactivePersistentObject$getCciaLogfile()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="getCciaMeta"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-getCciaMeta'><code>cecelia::ReactivePersistentObject$getCciaMeta()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="getCciaName"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-getCciaName'><code>cecelia::ReactivePersistentObject$getCciaName()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="getCciaType"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-getCciaType'><code>cecelia::ReactivePersistentObject$getCciaType()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="getUID"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-getUID'><code>cecelia::ReactivePersistentObject$getUID()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="initialize"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-initialize'><code>cecelia::ReactivePersistentObject$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="isTransaction"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-isTransaction'><code>cecelia::ReactivePersistentObject$isTransaction()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="lockReleased"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-lockReleased'><code>cecelia::ReactivePersistentObject$lockReleased()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="logfile"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-logfile'><code>cecelia::ReactivePersistentObject$logfile()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="logfileNames"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-logfileNames'><code>cecelia::ReactivePersistentObject$logfileNames()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="moduleFunParams"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-moduleFunParams'><code>cecelia::ReactivePersistentObject$moduleFunParams()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="objFiles"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-objFiles'><code>cecelia::ReactivePersistentObject$objFiles()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="persistentObjectDirectory"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-persistentObjectDirectory'><code>cecelia::ReactivePersistentObject$persistentObjectDirectory()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="persistentObjectDirectoryFile"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-persistentObjectDirectoryFile'><code>cecelia::ReactivePersistentObject$persistentObjectDirectoryFile()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="runTask"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-runTask'><code>cecelia::ReactivePersistentObject$runTask()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="saveModuleFunParams"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-saveModuleFunParams'><code>cecelia::ReactivePersistentObject$saveModuleFunParams()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="setCciaAttr"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-setCciaAttr'><code>cecelia::ReactivePersistentObject$setCciaAttr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="setCciaClass"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-setCciaClass'><code>cecelia::ReactivePersistentObject$setCciaClass()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="setCciaLogfile"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-setCciaLogfile'><code>cecelia::ReactivePersistentObject$setCciaLogfile()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="setCciaMeta"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-setCciaMeta'><code>cecelia::ReactivePersistentObject$setCciaMeta()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="setCciaName"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-setCciaName'><code>cecelia::ReactivePersistentObject$setCciaName()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="setCciaType"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-setCciaType'><code>cecelia::ReactivePersistentObject$setCciaType()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="setUID"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-setUID'><code>cecelia::ReactivePersistentObject$setUID()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="taskConf"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-taskConf'><code>cecelia::ReactivePersistentObject$taskConf()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="version"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-version'><code>cecelia::ReactivePersistentObject$version()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObjectSet" data-id="addAttrForCciaObjects"><a href='../../cecelia/html/ReactivePersistentObjectSet.html#method-ReactivePersistentObjectSet-addAttrForCciaObjects'><code>cecelia::ReactivePersistentObjectSet$addAttrForCciaObjects()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObjectSet" data-id="addCciaObjects"><a href='../../cecelia/html/ReactivePersistentObjectSet.html#method-ReactivePersistentObjectSet-addCciaObjects'><code>cecelia::ReactivePersistentObjectSet$addCciaObjects()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObjectSet" data-id="cciaObjectByUID"><a href='../../cecelia/html/ReactivePersistentObjectSet.html#method-ReactivePersistentObjectSet-cciaObjectByUID'><code>cecelia::ReactivePersistentObjectSet$cciaObjectByUID()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObjectSet" data-id="cciaObjectMetas"><a href='../../cecelia/html/ReactivePersistentObjectSet.html#method-ReactivePersistentObjectSet-cciaObjectMetas'><code>cecelia::ReactivePersistentObjectSet$cciaObjectMetas()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObjectSet" data-id="cciaObjectNames"><a href='../../cecelia/html/ReactivePersistentObjectSet.html#method-ReactivePersistentObjectSet-cciaObjectNames'><code>cecelia::ReactivePersistentObjectSet$cciaObjectNames()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObjectSet" data-id="cciaObjectUIDs"><a href='../../cecelia/html/ReactivePersistentObjectSet.html#method-ReactivePersistentObjectSet-cciaObjectUIDs'><code>cecelia::ReactivePersistentObjectSet$cciaObjectUIDs()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObjectSet" data-id="cciaObjects"><a href='../../cecelia/html/ReactivePersistentObjectSet.html#method-ReactivePersistentObjectSet-cciaObjects'><code>cecelia::ReactivePersistentObjectSet$cciaObjects()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObjectSet" data-id="cciaObjectsAttrNames"><a href='../../cecelia/html/ReactivePersistentObjectSet.html#method-ReactivePersistentObjectSet-cciaObjectsAttrNames'><code>cecelia::ReactivePersistentObjectSet$cciaObjectsAttrNames()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObjectSet" data-id="cciaObjectsByAttr"><a href='../../cecelia/html/ReactivePersistentObjectSet.html#method-ReactivePersistentObjectSet-cciaObjectsByAttr'><code>cecelia::ReactivePersistentObjectSet$cciaObjectsByAttr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObjectSet" data-id="cciaObjectsByUIDs"><a href='../../cecelia/html/ReactivePersistentObjectSet.html#method-ReactivePersistentObjectSet-cciaObjectsByUIDs'><code>cecelia::ReactivePersistentObjectSet$cciaObjectsByUIDs()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObjectSet" data-id="delAttrForCciaObjects"><a href='../../cecelia/html/ReactivePersistentObjectSet.html#method-ReactivePersistentObjectSet-delAttrForCciaObjects'><code>cecelia::ReactivePersistentObjectSet$delAttrForCciaObjects()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObjectSet" data-id="editAttrForCciaObjects"><a href='../../cecelia/html/ReactivePersistentObjectSet.html#method-ReactivePersistentObjectSet-editAttrForCciaObjects'><code>cecelia::ReactivePersistentObjectSet$editAttrForCciaObjects()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObjectSet" data-id="getCciaObjectUIDs"><a href='../../cecelia/html/ReactivePersistentObjectSet.html#method-ReactivePersistentObjectSet-getCciaObjectUIDs'><code>cecelia::ReactivePersistentObjectSet$getCciaObjectUIDs()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObjectSet" data-id="removeCciaObjectByUID"><a href='../../cecelia/html/ReactivePersistentObjectSet.html#method-ReactivePersistentObjectSet-removeCciaObjectByUID'><code>cecelia::ReactivePersistentObjectSet$removeCciaObjectByUID()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObjectSet" data-id="removeCciaObjects"><a href='../../cecelia/html/ReactivePersistentObjectSet.html#method-ReactivePersistentObjectSet-removeCciaObjects'><code>cecelia::ReactivePersistentObjectSet$removeCciaObjects()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObjectSet" data-id="retrieveState"><a href='../../cecelia/html/ReactivePersistentObjectSet.html#method-ReactivePersistentObjectSet-retrieveState'><code>cecelia::ReactivePersistentObjectSet$retrieveState()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObjectSet" data-id="runTasks"><a href='../../cecelia/html/ReactivePersistentObjectSet.html#method-ReactivePersistentObjectSet-runTasks'><code>cecelia::ReactivePersistentObjectSet$runTasks()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObjectSet" data-id="saveState"><a href='../../cecelia/html/ReactivePersistentObjectSet.html#method-ReactivePersistentObjectSet-saveState'><code>cecelia::ReactivePersistentObjectSet$saveState()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObjectSet" data-id="setCciaObjectUIDs"><a href='../../cecelia/html/ReactivePersistentObjectSet.html#method-ReactivePersistentObjectSet-setCciaObjectUIDs'><code>cecelia::ReactivePersistentObjectSet$setCciaObjectUIDs()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObjectSet" data-id="summary"><a href='../../cecelia/html/ReactivePersistentObjectSet.html#method-ReactivePersistentObjectSet-summary'><code>cecelia::ReactivePersistentObjectSet$summary()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactivePersistentObjectCollection-addCciaObjectsToSet"></a>}}
\if{latex}{\out{\hypertarget{method-ReactivePersistentObjectCollection-addCciaObjectsToSet}{}}}
\subsection{Method \code{addCciaObjectsToSet()}}{
uID by Set name


Add objects to set
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactivePersistentObjectCollection$addCciaObjectsToSet(setName, objectsToAdd)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{setName}}{character of Set name}

\item{\code{objectsToAdd}}{list of ReactivePersistentObjectSet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactivePersistentObjectCollection-addCciaObjectSet"></a>}}
\if{latex}{\out{\hypertarget{method-ReactivePersistentObjectCollection-addCciaObjectSet}{}}}
\subsection{Method \code{addCciaObjectSet()}}{
Add new Set
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactivePersistentObjectCollection$addCciaObjectSet(newSet)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{newSet}}{ReactivePersistentObjectSet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactivePersistentObjectCollection-cciaObjectSetByName"></a>}}
\if{latex}{\out{\hypertarget{method-ReactivePersistentObjectCollection-cciaObjectSetByName}{}}}
\subsection{Method \code{cciaObjectSetByName()}}{
Objects by Name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactivePersistentObjectCollection$cciaObjectSetByName(setName)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{setName}}{character of Set name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactivePersistentObjectCollection-cciaNodeObjectByUID"></a>}}
\if{latex}{\out{\hypertarget{method-ReactivePersistentObjectCollection-cciaNodeObjectByUID}{}}}
\subsection{Method \code{cciaNodeObjectByUID()}}{
ccia node object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactivePersistentObjectCollection$cciaNodeObjectByUID(uID)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{uID}}{character of unique ID}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactivePersistentObjectCollection-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ReactivePersistentObjectCollection-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactivePersistentObjectCollection$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
