% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskCore.R
\name{TaskCore}
\alias{TaskCore}
\title{Core class to execute module tasks}
\description{
Core class for task functions
}
\examples{
TODO
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TaskCore-new}{\code{TaskCore$new()}}
\item \href{#method-TaskCore-runTasks}{\code{TaskCore$runTasks()}}
\item \href{#method-TaskCore-setFunParams}{\code{TaskCore$setFunParams()}}
\item \href{#method-TaskCore-funParams}{\code{TaskCore$funParams()}}
\item \href{#method-TaskCore-taskEnv}{\code{TaskCore$taskEnv()}}
\item \href{#method-TaskCore-callingEnv}{\code{TaskCore$callingEnv()}}
\item \href{#method-TaskCore-envParams}{\code{TaskCore$envParams()}}
\item \href{#method-TaskCore-globalParams}{\code{TaskCore$globalParams()}}
\item \href{#method-TaskCore-utilsParams}{\code{TaskCore$utilsParams()}}
\item \href{#method-TaskCore-clone}{\code{TaskCore$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaskCore-new"></a>}}
\if{latex}{\out{\hypertarget{method-TaskCore-new}{}}}
\subsection{Method \code{new()}}{
Init
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaskCore$new()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{moduleName}}{character of module name}

\item{\code{funName}}{character of functions name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaskCore-runTasks"></a>}}
\if{latex}{\out{\hypertarget{method-TaskCore-runTasks}{}}}
\subsection{Method \code{runTasks()}}{
Call other task processes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaskCore$runTasks(
  tasks,
  taskConf = NULL,
  funParams = list(),
  runLocal = FALSE,
  callingEnv = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tasks}}{list of character for task functions}

\item{\code{taskConf}}{list of list of character for task config}

\item{\code{funParams}}{list of generic for task parameters}

\item{\code{runLocal}}{boolean to run task local}

\item{\code{callingEnv}}{character for calling environment}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaskCore-setFunParams"></a>}}
\if{latex}{\out{\hypertarget{method-TaskCore-setFunParams}{}}}
\subsection{Method \code{setFunParams()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaskCore$setFunParams(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaskCore-funParams"></a>}}
\if{latex}{\out{\hypertarget{method-TaskCore-funParams}{}}}
\subsection{Method \code{funParams()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaskCore$funParams()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaskCore-taskEnv"></a>}}
\if{latex}{\out{\hypertarget{method-TaskCore-taskEnv}{}}}
\subsection{Method \code{taskEnv()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaskCore$taskEnv()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaskCore-callingEnv"></a>}}
\if{latex}{\out{\hypertarget{method-TaskCore-callingEnv}{}}}
\subsection{Method \code{callingEnv()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaskCore$callingEnv()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaskCore-envParams"></a>}}
\if{latex}{\out{\hypertarget{method-TaskCore-envParams}{}}}
\subsection{Method \code{envParams()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaskCore$envParams(taskEnv = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaskCore-globalParams"></a>}}
\if{latex}{\out{\hypertarget{method-TaskCore-globalParams}{}}}
\subsection{Method \code{globalParams()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaskCore$globalParams()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaskCore-utilsParams"></a>}}
\if{latex}{\out{\hypertarget{method-TaskCore-utilsParams}{}}}
\subsection{Method \code{utilsParams()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaskCore$utilsParams()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaskCore-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TaskCore-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaskCore$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
