% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cciaImageCollection.R
\name{CciaImageCollection}
\alias{CciaImageCollection}
\title{Reactive image object collection}
\description{
Reactive image object collection
}
\examples{
TODO
}
\section{Super classes}{
\code{\link[ceceliaCore:ReactiveObject]{ceceliaCore::ReactiveObject}} -> \code{\link[ceceliaCore:ReactivePersistentObject]{ceceliaCore::ReactivePersistentObject}} -> \code{\link[ceceliaCore:ReactivePersistentObjectSet]{ceceliaCore::ReactivePersistentObjectSet}} -> \code{\link[ceceliaCore:ReactivePersistentObjectCollection]{ceceliaCore::ReactivePersistentObjectCollection}} -> \code{CciaImageCollection}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CciaImageCollection-new}{\code{CciaImageCollection$new()}}
\item \href{#method-CciaImageCollection-clFiles}{\code{CciaImageCollection$clFiles()}}
\item \href{#method-CciaImageCollection-pixclFiles}{\code{CciaImageCollection$pixclFiles()}}
\item \href{#method-CciaImageCollection-clone}{\code{CciaImageCollection$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="ReactiveObject" data-id="reactive"><a href='../../ceceliaCore/html/ReactiveObject.html#method-ReactiveObject-reactive'><code>ceceliaCore::ReactiveObject$reactive()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="ReactiveObject" data-id="update"><a href='../../ceceliaCore/html/ReactiveObject.html#method-ReactiveObject-update'><code>ceceliaCore::ReactiveObject$update()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="ReactivePersistentObject" data-id="addCciaAttr"><a href='../../ceceliaCore/html/ReactivePersistentObject.html#method-ReactivePersistentObject-addCciaAttr'><code>ceceliaCore::ReactivePersistentObject$addCciaAttr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="ReactivePersistentObject" data-id="delCciaAttr"><a href='../../ceceliaCore/html/ReactivePersistentObject.html#method-ReactivePersistentObject-delCciaAttr'><code>ceceliaCore::ReactivePersistentObject$delCciaAttr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="ReactivePersistentObject" data-id="deleteObjectDirectory"><a href='../../ceceliaCore/html/ReactivePersistentObject.html#method-ReactivePersistentObject-deleteObjectDirectory'><code>ceceliaCore::ReactivePersistentObject$deleteObjectDirectory()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="ReactivePersistentObject" data-id="editCciaAttr"><a href='../../ceceliaCore/html/ReactivePersistentObject.html#method-ReactivePersistentObject-editCciaAttr'><code>ceceliaCore::ReactivePersistentObject$editCciaAttr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="ReactivePersistentObject" data-id="finalize"><a href='../../ceceliaCore/html/ReactivePersistentObject.html#method-ReactivePersistentObject-finalize'><code>ceceliaCore::ReactivePersistentObject$finalize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="ReactivePersistentObject" data-id="getCciaAttr"><a href='../../ceceliaCore/html/ReactivePersistentObject.html#method-ReactivePersistentObject-getCciaAttr'><code>ceceliaCore::ReactivePersistentObject$getCciaAttr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="ReactivePersistentObject" data-id="getCciaClass"><a href='../../ceceliaCore/html/ReactivePersistentObject.html#method-ReactivePersistentObject-getCciaClass'><code>ceceliaCore::ReactivePersistentObject$getCciaClass()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="ReactivePersistentObject" data-id="getCciaLogfile"><a href='../../ceceliaCore/html/ReactivePersistentObject.html#method-ReactivePersistentObject-getCciaLogfile'><code>ceceliaCore::ReactivePersistentObject$getCciaLogfile()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="ReactivePersistentObject" data-id="getCciaMeta"><a href='../../ceceliaCore/html/ReactivePersistentObject.html#method-ReactivePersistentObject-getCciaMeta'><code>ceceliaCore::ReactivePersistentObject$getCciaMeta()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="ReactivePersistentObject" data-id="getCciaName"><a href='../../ceceliaCore/html/ReactivePersistentObject.html#method-ReactivePersistentObject-getCciaName'><code>ceceliaCore::ReactivePersistentObject$getCciaName()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="ReactivePersistentObject" data-id="getCciaType"><a href='../../ceceliaCore/html/ReactivePersistentObject.html#method-ReactivePersistentObject-getCciaType'><code>ceceliaCore::ReactivePersistentObject$getCciaType()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="ReactivePersistentObject" data-id="getUID"><a href='../../ceceliaCore/html/ReactivePersistentObject.html#method-ReactivePersistentObject-getUID'><code>ceceliaCore::ReactivePersistentObject$getUID()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="ReactivePersistentObject" data-id="isTransaction"><a href='../../ceceliaCore/html/ReactivePersistentObject.html#method-ReactivePersistentObject-isTransaction'><code>ceceliaCore::ReactivePersistentObject$isTransaction()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="ReactivePersistentObject" data-id="lockReleased"><a href='../../ceceliaCore/html/ReactivePersistentObject.html#method-ReactivePersistentObject-lockReleased'><code>ceceliaCore::ReactivePersistentObject$lockReleased()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="ReactivePersistentObject" data-id="logfile"><a href='../../ceceliaCore/html/ReactivePersistentObject.html#method-ReactivePersistentObject-logfile'><code>ceceliaCore::ReactivePersistentObject$logfile()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="ReactivePersistentObject" data-id="logfileNames"><a href='../../ceceliaCore/html/ReactivePersistentObject.html#method-ReactivePersistentObject-logfileNames'><code>ceceliaCore::ReactivePersistentObject$logfileNames()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="ReactivePersistentObject" data-id="moduleFunParams"><a href='../../ceceliaCore/html/ReactivePersistentObject.html#method-ReactivePersistentObject-moduleFunParams'><code>ceceliaCore::ReactivePersistentObject$moduleFunParams()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="ReactivePersistentObject" data-id="objFiles"><a href='../../ceceliaCore/html/ReactivePersistentObject.html#method-ReactivePersistentObject-objFiles'><code>ceceliaCore::ReactivePersistentObject$objFiles()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="ReactivePersistentObject" data-id="persistentObjectDirectory"><a href='../../ceceliaCore/html/ReactivePersistentObject.html#method-ReactivePersistentObject-persistentObjectDirectory'><code>ceceliaCore::ReactivePersistentObject$persistentObjectDirectory()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="ReactivePersistentObject" data-id="persistentObjectDirectoryFile"><a href='../../ceceliaCore/html/ReactivePersistentObject.html#method-ReactivePersistentObject-persistentObjectDirectoryFile'><code>ceceliaCore::ReactivePersistentObject$persistentObjectDirectoryFile()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="ReactivePersistentObject" data-id="runTask"><a href='../../ceceliaCore/html/ReactivePersistentObject.html#method-ReactivePersistentObject-runTask'><code>ceceliaCore::ReactivePersistentObject$runTask()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="ReactivePersistentObject" data-id="saveModuleFunParams"><a href='../../ceceliaCore/html/ReactivePersistentObject.html#method-ReactivePersistentObject-saveModuleFunParams'><code>ceceliaCore::ReactivePersistentObject$saveModuleFunParams()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="ReactivePersistentObject" data-id="setCciaAttr"><a href='../../ceceliaCore/html/ReactivePersistentObject.html#method-ReactivePersistentObject-setCciaAttr'><code>ceceliaCore::ReactivePersistentObject$setCciaAttr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="ReactivePersistentObject" data-id="setCciaClass"><a href='../../ceceliaCore/html/ReactivePersistentObject.html#method-ReactivePersistentObject-setCciaClass'><code>ceceliaCore::ReactivePersistentObject$setCciaClass()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="ReactivePersistentObject" data-id="setCciaLogfile"><a href='../../ceceliaCore/html/ReactivePersistentObject.html#method-ReactivePersistentObject-setCciaLogfile'><code>ceceliaCore::ReactivePersistentObject$setCciaLogfile()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="ReactivePersistentObject" data-id="setCciaMeta"><a href='../../ceceliaCore/html/ReactivePersistentObject.html#method-ReactivePersistentObject-setCciaMeta'><code>ceceliaCore::ReactivePersistentObject$setCciaMeta()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="ReactivePersistentObject" data-id="setCciaName"><a href='../../ceceliaCore/html/ReactivePersistentObject.html#method-ReactivePersistentObject-setCciaName'><code>ceceliaCore::ReactivePersistentObject$setCciaName()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="ReactivePersistentObject" data-id="setCciaType"><a href='../../ceceliaCore/html/ReactivePersistentObject.html#method-ReactivePersistentObject-setCciaType'><code>ceceliaCore::ReactivePersistentObject$setCciaType()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="ReactivePersistentObject" data-id="setUID"><a href='../../ceceliaCore/html/ReactivePersistentObject.html#method-ReactivePersistentObject-setUID'><code>ceceliaCore::ReactivePersistentObject$setUID()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="ReactivePersistentObject" data-id="taskConf"><a href='../../ceceliaCore/html/ReactivePersistentObject.html#method-ReactivePersistentObject-taskConf'><code>ceceliaCore::ReactivePersistentObject$taskConf()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="ReactivePersistentObject" data-id="version"><a href='../../ceceliaCore/html/ReactivePersistentObject.html#method-ReactivePersistentObject-version'><code>ceceliaCore::ReactivePersistentObject$version()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="ReactivePersistentObjectSet" data-id="addAttrForCciaObjects"><a href='../../ceceliaCore/html/ReactivePersistentObjectSet.html#method-ReactivePersistentObjectSet-addAttrForCciaObjects'><code>ceceliaCore::ReactivePersistentObjectSet$addAttrForCciaObjects()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="ReactivePersistentObjectSet" data-id="addCciaObjects"><a href='../../ceceliaCore/html/ReactivePersistentObjectSet.html#method-ReactivePersistentObjectSet-addCciaObjects'><code>ceceliaCore::ReactivePersistentObjectSet$addCciaObjects()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="ReactivePersistentObjectSet" data-id="cciaObjectByUID"><a href='../../ceceliaCore/html/ReactivePersistentObjectSet.html#method-ReactivePersistentObjectSet-cciaObjectByUID'><code>ceceliaCore::ReactivePersistentObjectSet$cciaObjectByUID()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="ReactivePersistentObjectSet" data-id="cciaObjectMetas"><a href='../../ceceliaCore/html/ReactivePersistentObjectSet.html#method-ReactivePersistentObjectSet-cciaObjectMetas'><code>ceceliaCore::ReactivePersistentObjectSet$cciaObjectMetas()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="ReactivePersistentObjectSet" data-id="cciaObjectNames"><a href='../../ceceliaCore/html/ReactivePersistentObjectSet.html#method-ReactivePersistentObjectSet-cciaObjectNames'><code>ceceliaCore::ReactivePersistentObjectSet$cciaObjectNames()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="ReactivePersistentObjectSet" data-id="cciaObjectUIDs"><a href='../../ceceliaCore/html/ReactivePersistentObjectSet.html#method-ReactivePersistentObjectSet-cciaObjectUIDs'><code>ceceliaCore::ReactivePersistentObjectSet$cciaObjectUIDs()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="ReactivePersistentObjectSet" data-id="cciaObjects"><a href='../../ceceliaCore/html/ReactivePersistentObjectSet.html#method-ReactivePersistentObjectSet-cciaObjects'><code>ceceliaCore::ReactivePersistentObjectSet$cciaObjects()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="ReactivePersistentObjectSet" data-id="cciaObjectsAttrNames"><a href='../../ceceliaCore/html/ReactivePersistentObjectSet.html#method-ReactivePersistentObjectSet-cciaObjectsAttrNames'><code>ceceliaCore::ReactivePersistentObjectSet$cciaObjectsAttrNames()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="ReactivePersistentObjectSet" data-id="cciaObjectsByAttr"><a href='../../ceceliaCore/html/ReactivePersistentObjectSet.html#method-ReactivePersistentObjectSet-cciaObjectsByAttr'><code>ceceliaCore::ReactivePersistentObjectSet$cciaObjectsByAttr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="ReactivePersistentObjectSet" data-id="cciaObjectsByUIDs"><a href='../../ceceliaCore/html/ReactivePersistentObjectSet.html#method-ReactivePersistentObjectSet-cciaObjectsByUIDs'><code>ceceliaCore::ReactivePersistentObjectSet$cciaObjectsByUIDs()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="ReactivePersistentObjectSet" data-id="delAttrForCciaObjects"><a href='../../ceceliaCore/html/ReactivePersistentObjectSet.html#method-ReactivePersistentObjectSet-delAttrForCciaObjects'><code>ceceliaCore::ReactivePersistentObjectSet$delAttrForCciaObjects()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="ReactivePersistentObjectSet" data-id="editAttrForCciaObjects"><a href='../../ceceliaCore/html/ReactivePersistentObjectSet.html#method-ReactivePersistentObjectSet-editAttrForCciaObjects'><code>ceceliaCore::ReactivePersistentObjectSet$editAttrForCciaObjects()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="ReactivePersistentObjectSet" data-id="getCciaObjectUIDs"><a href='../../ceceliaCore/html/ReactivePersistentObjectSet.html#method-ReactivePersistentObjectSet-getCciaObjectUIDs'><code>ceceliaCore::ReactivePersistentObjectSet$getCciaObjectUIDs()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="ReactivePersistentObjectSet" data-id="removeCciaObjectByUID"><a href='../../ceceliaCore/html/ReactivePersistentObjectSet.html#method-ReactivePersistentObjectSet-removeCciaObjectByUID'><code>ceceliaCore::ReactivePersistentObjectSet$removeCciaObjectByUID()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="ReactivePersistentObjectSet" data-id="removeCciaObjects"><a href='../../ceceliaCore/html/ReactivePersistentObjectSet.html#method-ReactivePersistentObjectSet-removeCciaObjects'><code>ceceliaCore::ReactivePersistentObjectSet$removeCciaObjects()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="ReactivePersistentObjectSet" data-id="retrieveState"><a href='../../ceceliaCore/html/ReactivePersistentObjectSet.html#method-ReactivePersistentObjectSet-retrieveState'><code>ceceliaCore::ReactivePersistentObjectSet$retrieveState()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="ReactivePersistentObjectSet" data-id="runTasks"><a href='../../ceceliaCore/html/ReactivePersistentObjectSet.html#method-ReactivePersistentObjectSet-runTasks'><code>ceceliaCore::ReactivePersistentObjectSet$runTasks()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="ReactivePersistentObjectSet" data-id="saveState"><a href='../../ceceliaCore/html/ReactivePersistentObjectSet.html#method-ReactivePersistentObjectSet-saveState'><code>ceceliaCore::ReactivePersistentObjectSet$saveState()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="ReactivePersistentObjectSet" data-id="setCciaObjectUIDs"><a href='../../ceceliaCore/html/ReactivePersistentObjectSet.html#method-ReactivePersistentObjectSet-setCciaObjectUIDs'><code>ceceliaCore::ReactivePersistentObjectSet$setCciaObjectUIDs()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="ReactivePersistentObjectSet" data-id="summary"><a href='../../ceceliaCore/html/ReactivePersistentObjectSet.html#method-ReactivePersistentObjectSet-summary'><code>ceceliaCore::ReactivePersistentObjectSet$summary()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="ReactivePersistentObjectCollection" data-id="addCciaObjectSet"><a href='../../ceceliaCore/html/ReactivePersistentObjectCollection.html#method-ReactivePersistentObjectCollection-addCciaObjectSet'><code>ceceliaCore::ReactivePersistentObjectCollection$addCciaObjectSet()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="ReactivePersistentObjectCollection" data-id="addCciaObjectsToSet"><a href='../../ceceliaCore/html/ReactivePersistentObjectCollection.html#method-ReactivePersistentObjectCollection-addCciaObjectsToSet'><code>ceceliaCore::ReactivePersistentObjectCollection$addCciaObjectsToSet()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="ReactivePersistentObjectCollection" data-id="cciaNodeObjectByUID"><a href='../../ceceliaCore/html/ReactivePersistentObjectCollection.html#method-ReactivePersistentObjectCollection-cciaNodeObjectByUID'><code>ceceliaCore::ReactivePersistentObjectCollection$cciaNodeObjectByUID()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="ReactivePersistentObjectCollection" data-id="cciaObjectSetByName"><a href='../../ceceliaCore/html/ReactivePersistentObjectCollection.html#method-ReactivePersistentObjectCollection-cciaObjectSetByName'><code>ceceliaCore::ReactivePersistentObjectCollection$cciaObjectSetByName()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImageCollection-new"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImageCollection-new}{}}}
\subsection{Method \code{new()}}{
Init
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImageCollection$new(stateFile, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{stateFile}}{character with file path to state file}

\item{\code{...}}{passed to super$initialize}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImageCollection-clFiles"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImageCollection-clFiles}{}}}
\subsection{Method \code{clFiles()}}{
Classifier files
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImageCollection$clFiles(clType, fullPath = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{clType}}{character of classifier type}

\item{\code{fullPath}}{boolean to return full path}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImageCollection-pixclFiles"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImageCollection-pixclFiles}{}}}
\subsection{Method \code{pixclFiles()}}{
Pixel classifier files
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImageCollection$pixclFiles(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{passed to self$clFiles}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CciaImageCollection-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CciaImageCollection-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CciaImageCollection$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
