% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multifileLabelPopUtils.R
\name{MultifileLabelPopUtils}
\alias{MultifileLabelPopUtils}
\title{Population utility that combine multiple label properties into one
population interface. I only used this for live cell tracking so far where
every base population is a single adata-file.}
\description{
Multi file label populations
}
\examples{
TODO
}
\section{Super classes}{
\code{\link[ceceliaCore:ReactiveObject]{ceceliaCore::ReactiveObject}} -> \code{\link[ceceliaCore:PopulationUtils]{ceceliaCore::PopulationUtils}} -> \code{MultifileLabelPopUtils}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MultifileLabelPopUtils-new}{\code{MultifileLabelPopUtils$new()}}
\item \href{#method-MultifileLabelPopUtils-defaultPops}{\code{MultifileLabelPopUtils$defaultPops()}}
\item \href{#method-MultifileLabelPopUtils-popDT}{\code{MultifileLabelPopUtils$popDT()}}
\item \href{#method-MultifileLabelPopUtils-save}{\code{MultifileLabelPopUtils$save()}}
\item \href{#method-MultifileLabelPopUtils-popPaths}{\code{MultifileLabelPopUtils$popPaths()}}
\item \href{#method-MultifileLabelPopUtils-popLabels}{\code{MultifileLabelPopUtils$popLabels()}}
\item \href{#method-MultifileLabelPopUtils-clone}{\code{MultifileLabelPopUtils$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="ReactiveObject" data-id="reactive"><a href='../../ceceliaCore/html/ReactiveObject.html#method-ReactiveObject-reactive'><code>ceceliaCore::ReactiveObject$reactive()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="ReactiveObject" data-id="update"><a href='../../ceceliaCore/html/ReactiveObject.html#method-ReactiveObject-update'><code>ceceliaCore::ReactiveObject$update()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="PopulationUtils" data-id="getImChannelLimits"><a href='../../ceceliaCore/html/PopulationUtils.html#method-PopulationUtils-getImChannelLimits'><code>ceceliaCore::PopulationUtils$getImChannelLimits()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="PopulationUtils" data-id="getImChannels"><a href='../../ceceliaCore/html/PopulationUtils.html#method-PopulationUtils-getImChannels'><code>ceceliaCore::PopulationUtils$getImChannels()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="PopulationUtils" data-id="getPopObj"><a href='../../ceceliaCore/html/PopulationUtils.html#method-PopulationUtils-getPopObj'><code>ceceliaCore::PopulationUtils$getPopObj()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="PopulationUtils" data-id="getPopObjFilepath"><a href='../../ceceliaCore/html/PopulationUtils.html#method-PopulationUtils-getPopObjFilepath'><code>ceceliaCore::PopulationUtils$getPopObjFilepath()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="PopulationUtils" data-id="getValueNames"><a href='../../ceceliaCore/html/PopulationUtils.html#method-PopulationUtils-getValueNames'><code>ceceliaCore::PopulationUtils$getValueNames()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="PopulationUtils" data-id="isLabelPropsStore"><a href='../../ceceliaCore/html/PopulationUtils.html#method-PopulationUtils-isLabelPropsStore'><code>ceceliaCore::PopulationUtils$isLabelPropsStore()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="PopulationUtils" data-id="savePops"><a href='../../ceceliaCore/html/PopulationUtils.html#method-PopulationUtils-savePops'><code>ceceliaCore::PopulationUtils$savePops()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ceceliaCore" data-topic="PopulationUtils" data-id="unlinkPops"><a href='../../ceceliaCore/html/PopulationUtils.html#method-PopulationUtils-unlinkPops'><code>ceceliaCore::PopulationUtils$unlinkPops()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultifileLabelPopUtils-new"></a>}}
\if{latex}{\out{\hypertarget{method-MultifileLabelPopUtils-new}{}}}
\subsection{Method \code{new()}}{
Init
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultifileLabelPopUtils$new(taskDir, valueNames, imChannels)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{taskDir}}{character for task directory}

\item{\code{valueNames}}{list of character for value names}

\item{\code{imChannels}}{list of character for channel names}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultifileLabelPopUtils-defaultPops"></a>}}
\if{latex}{\out{\hypertarget{method-MultifileLabelPopUtils-defaultPops}{}}}
\subsection{Method \code{defaultPops()}}{
default pops
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultifileLabelPopUtils$defaultPops()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultifileLabelPopUtils-popDT"></a>}}
\if{latex}{\out{\hypertarget{method-MultifileLabelPopUtils-popDT}{}}}
\subsection{Method \code{popDT()}}{
Population data.table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultifileLabelPopUtils$popDT(
  pops = NULL,
  cols = NULL,
  dropNA = FALSE,
  dropPop = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pops}}{list of character for populations}

\item{\code{cols}}{list of character for columns}

\item{\code{dropNA}}{boolean to drop NA}

\item{\code{dropPop}}{boolean to drop population}

\item{\code{copyDT}}{boolean to copy data.table}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultifileLabelPopUtils-save"></a>}}
\if{latex}{\out{\hypertarget{method-MultifileLabelPopUtils-save}{}}}
\subsection{Method \code{save()}}{
Save anndata back to disk
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultifileLabelPopUtils$save()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultifileLabelPopUtils-popPaths"></a>}}
\if{latex}{\out{\hypertarget{method-MultifileLabelPopUtils-popPaths}{}}}
\subsection{Method \code{popPaths()}}{
Populations paths
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultifileLabelPopUtils$popPaths(includeRoot = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{includeRoot}}{boolean to include root}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultifileLabelPopUtils-popLabels"></a>}}
\if{latex}{\out{\hypertarget{method-MultifileLabelPopUtils-popLabels}{}}}
\subsection{Method \code{popLabels()}}{
Pop labels
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultifileLabelPopUtils$popLabels()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultifileLabelPopUtils-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MultifileLabelPopUtils-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultifileLabelPopUtils$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
