% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactivePersistentObjectSet.R
\name{ReactivePersistentObjectSet}
\alias{ReactivePersistentObjectSet}
\title{Set of reactive persistent objects.}
\description{
Set of persistent reactive objects
}
\examples{
TODO
}
\section{Super classes}{
\code{\link[cecelia:ReactiveObject]{cecelia::ReactiveObject}} -> \code{\link[cecelia:ReactivePersistentObject]{cecelia::ReactivePersistentObject}} -> \code{ReactivePersistentObjectSet}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ReactivePersistentObjectSet-addCciaObjects}{\code{ReactivePersistentObjectSet$addCciaObjects()}}
\item \href{#method-ReactivePersistentObjectSet-removeCciaObjects}{\code{ReactivePersistentObjectSet$removeCciaObjects()}}
\item \href{#method-ReactivePersistentObjectSet-removeCciaObjectByUID}{\code{ReactivePersistentObjectSet$removeCciaObjectByUID()}}
\item \href{#method-ReactivePersistentObjectSet-cciaObjectsAttrNames}{\code{ReactivePersistentObjectSet$cciaObjectsAttrNames()}}
\item \href{#method-ReactivePersistentObjectSet-addAttrForCciaObjects}{\code{ReactivePersistentObjectSet$addAttrForCciaObjects()}}
\item \href{#method-ReactivePersistentObjectSet-editAttrForCciaObjects}{\code{ReactivePersistentObjectSet$editAttrForCciaObjects()}}
\item \href{#method-ReactivePersistentObjectSet-delAttrForCciaObjects}{\code{ReactivePersistentObjectSet$delAttrForCciaObjects()}}
\item \href{#method-ReactivePersistentObjectSet-saveState}{\code{ReactivePersistentObjectSet$saveState()}}
\item \href{#method-ReactivePersistentObjectSet-retrieveState}{\code{ReactivePersistentObjectSet$retrieveState()}}
\item \href{#method-ReactivePersistentObjectSet-summary}{\code{ReactivePersistentObjectSet$summary()}}
\item \href{#method-ReactivePersistentObjectSet-runTasks}{\code{ReactivePersistentObjectSet$runTasks()}}
\item \href{#method-ReactivePersistentObjectSet-cciaObjects}{\code{ReactivePersistentObjectSet$cciaObjects()}}
\item \href{#method-ReactivePersistentObjectSet-cciaObjectsByUIDs}{\code{ReactivePersistentObjectSet$cciaObjectsByUIDs()}}
\item \href{#method-ReactivePersistentObjectSet-cciaObjectByUID}{\code{ReactivePersistentObjectSet$cciaObjectByUID()}}
\item \href{#method-ReactivePersistentObjectSet-cciaObjectsByAttr}{\code{ReactivePersistentObjectSet$cciaObjectsByAttr()}}
\item \href{#method-ReactivePersistentObjectSet-setCciaObjectUIDs}{\code{ReactivePersistentObjectSet$setCciaObjectUIDs()}}
\item \href{#method-ReactivePersistentObjectSet-getCciaObjectUIDs}{\code{ReactivePersistentObjectSet$getCciaObjectUIDs()}}
\item \href{#method-ReactivePersistentObjectSet-cciaObjectUIDs}{\code{ReactivePersistentObjectSet$cciaObjectUIDs()}}
\item \href{#method-ReactivePersistentObjectSet-cciaObjectNames}{\code{ReactivePersistentObjectSet$cciaObjectNames()}}
\item \href{#method-ReactivePersistentObjectSet-cciaObjectMetas}{\code{ReactivePersistentObjectSet$cciaObjectMetas()}}
\item \href{#method-ReactivePersistentObjectSet-clone}{\code{ReactivePersistentObjectSet$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactiveObject" data-id="reactive"><a href='../../cecelia/html/ReactiveObject.html#method-ReactiveObject-reactive'><code>cecelia::ReactiveObject$reactive()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactiveObject" data-id="update"><a href='../../cecelia/html/ReactiveObject.html#method-ReactiveObject-update'><code>cecelia::ReactiveObject$update()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="addCciaAttr"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-addCciaAttr'><code>cecelia::ReactivePersistentObject$addCciaAttr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="delCciaAttr"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-delCciaAttr'><code>cecelia::ReactivePersistentObject$delCciaAttr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="deleteObjectDirectory"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-deleteObjectDirectory'><code>cecelia::ReactivePersistentObject$deleteObjectDirectory()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="editCciaAttr"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-editCciaAttr'><code>cecelia::ReactivePersistentObject$editCciaAttr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="finalize"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-finalize'><code>cecelia::ReactivePersistentObject$finalize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="getCciaAttr"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-getCciaAttr'><code>cecelia::ReactivePersistentObject$getCciaAttr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="getCciaClass"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-getCciaClass'><code>cecelia::ReactivePersistentObject$getCciaClass()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="getCciaLogfile"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-getCciaLogfile'><code>cecelia::ReactivePersistentObject$getCciaLogfile()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="getCciaMeta"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-getCciaMeta'><code>cecelia::ReactivePersistentObject$getCciaMeta()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="getCciaName"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-getCciaName'><code>cecelia::ReactivePersistentObject$getCciaName()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="getCciaType"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-getCciaType'><code>cecelia::ReactivePersistentObject$getCciaType()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="getUID"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-getUID'><code>cecelia::ReactivePersistentObject$getUID()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="initialize"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-initialize'><code>cecelia::ReactivePersistentObject$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="isTransaction"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-isTransaction'><code>cecelia::ReactivePersistentObject$isTransaction()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="lockReleased"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-lockReleased'><code>cecelia::ReactivePersistentObject$lockReleased()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="logfile"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-logfile'><code>cecelia::ReactivePersistentObject$logfile()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="logfileNames"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-logfileNames'><code>cecelia::ReactivePersistentObject$logfileNames()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="moduleFunParams"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-moduleFunParams'><code>cecelia::ReactivePersistentObject$moduleFunParams()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="objFiles"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-objFiles'><code>cecelia::ReactivePersistentObject$objFiles()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="persistentObjectDirectory"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-persistentObjectDirectory'><code>cecelia::ReactivePersistentObject$persistentObjectDirectory()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="persistentObjectDirectoryFile"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-persistentObjectDirectoryFile'><code>cecelia::ReactivePersistentObject$persistentObjectDirectoryFile()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="runTask"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-runTask'><code>cecelia::ReactivePersistentObject$runTask()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="saveModuleFunParams"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-saveModuleFunParams'><code>cecelia::ReactivePersistentObject$saveModuleFunParams()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="setCciaAttr"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-setCciaAttr'><code>cecelia::ReactivePersistentObject$setCciaAttr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="setCciaClass"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-setCciaClass'><code>cecelia::ReactivePersistentObject$setCciaClass()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="setCciaLogfile"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-setCciaLogfile'><code>cecelia::ReactivePersistentObject$setCciaLogfile()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="setCciaMeta"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-setCciaMeta'><code>cecelia::ReactivePersistentObject$setCciaMeta()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="setCciaName"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-setCciaName'><code>cecelia::ReactivePersistentObject$setCciaName()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="setCciaType"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-setCciaType'><code>cecelia::ReactivePersistentObject$setCciaType()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="setUID"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-setUID'><code>cecelia::ReactivePersistentObject$setUID()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="taskConf"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-taskConf'><code>cecelia::ReactivePersistentObject$taskConf()</code></a></span></li>
<li><span class="pkg-link" data-pkg="cecelia" data-topic="ReactivePersistentObject" data-id="version"><a href='../../cecelia/html/ReactivePersistentObject.html#method-ReactivePersistentObject-version'><code>cecelia::ReactivePersistentObject$version()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactivePersistentObjectSet-addCciaObjects"></a>}}
\if{latex}{\out{\hypertarget{method-ReactivePersistentObjectSet-addCciaObjects}{}}}
\subsection{Method \code{addCciaObjects()}}{
Param from objects


Check that parameter is a object list


Add objects
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactivePersistentObjectSet$addCciaObjects(
  cciaObjects,
  reset = FALSE,
  invalidate = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cciaObjects}}{list of ReactivePersistentObject}

\item{\code{cciaObjects}}{list of ReactivePersistentObject}

\item{\code{cciaObjects}}{list of ReactivePersistentObject}

\item{\code{reset}}{boolean to reset list}

\item{\code{invalidate}}{boolean to invalidate object}

\item{\code{param}}{character of param to get}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactivePersistentObjectSet-removeCciaObjects"></a>}}
\if{latex}{\out{\hypertarget{method-ReactivePersistentObjectSet-removeCciaObjects}{}}}
\subsection{Method \code{removeCciaObjects()}}{
Remove objects
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactivePersistentObjectSet$removeCciaObjects(
  cciaObjects = NULL,
  removeContent = TRUE,
  invalidate = TRUE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cciaObjects}}{list of ReactivePersistentObject}

\item{\code{removeContent}}{boolean to remove object content}

\item{\code{invalidate}}{boolean to invalidate object}

\item{\code{...}}{ReactivePersistentObject$deleteObjectDirectory}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactivePersistentObjectSet-removeCciaObjectByUID"></a>}}
\if{latex}{\out{\hypertarget{method-ReactivePersistentObjectSet-removeCciaObjectByUID}{}}}
\subsection{Method \code{removeCciaObjectByUID()}}{
Remove object by uID
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactivePersistentObjectSet$removeCciaObjectByUID(uID, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{uID}}{character unique ID}

\item{\code{...}}{self$removeCciaObjects}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactivePersistentObjectSet-cciaObjectsAttrNames"></a>}}
\if{latex}{\out{\hypertarget{method-ReactivePersistentObjectSet-cciaObjectsAttrNames}{}}}
\subsection{Method \code{cciaObjectsAttrNames()}}{
Attributes for objects
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactivePersistentObjectSet$cciaObjectsAttrNames(cciaObjects = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cciaObjects}}{list of ReactivePersistentObject}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactivePersistentObjectSet-addAttrForCciaObjects"></a>}}
\if{latex}{\out{\hypertarget{method-ReactivePersistentObjectSet-addAttrForCciaObjects}{}}}
\subsection{Method \code{addAttrForCciaObjects()}}{
Add attributes for objects
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactivePersistentObjectSet$addAttrForCciaObjects(
  objAttrName,
  objAttrVals = NULL,
  cciaObjects = NULL,
  invalidate = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{objAttrName}}{character for attribute name}

\item{\code{objAttrVals}}{character for attribute value}

\item{\code{cciaObjects}}{list of ReactivePersistentObject}

\item{\code{invalidate}}{boolean to invalidate object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactivePersistentObjectSet-editAttrForCciaObjects"></a>}}
\if{latex}{\out{\hypertarget{method-ReactivePersistentObjectSet-editAttrForCciaObjects}{}}}
\subsection{Method \code{editAttrForCciaObjects()}}{
Edit attributes for objects
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactivePersistentObjectSet$editAttrForCciaObjects(
  objAttrName,
  objAttrVals,
  cciaObjects = NULL,
  invalidate = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{objAttrName}}{character for attribute name}

\item{\code{objAttrVals}}{character for attribute value}

\item{\code{cciaObjects}}{list of ReactivePersistentObject}

\item{\code{invalidate}}{boolean to invalidate object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactivePersistentObjectSet-delAttrForCciaObjects"></a>}}
\if{latex}{\out{\hypertarget{method-ReactivePersistentObjectSet-delAttrForCciaObjects}{}}}
\subsection{Method \code{delAttrForCciaObjects()}}{
Delete attributes for objects
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactivePersistentObjectSet$delAttrForCciaObjects(
  objAttrName,
  cciaObjects = NULL,
  invalidate = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{objAttrName}}{character for attribute name}

\item{\code{cciaObjects}}{list of ReactivePersistentObject}

\item{\code{invalidate}}{boolean to invalidate object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactivePersistentObjectSet-saveState"></a>}}
\if{latex}{\out{\hypertarget{method-ReactivePersistentObjectSet-saveState}{}}}
\subsection{Method \code{saveState()}}{
Save objects
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactivePersistentObjectSet$saveState(includeChildren = TRUE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{includeChildren}}{boolean to include child nodes}

\item{\code{...}}{super$saveState & ReactivePersistentObject$saveState}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactivePersistentObjectSet-retrieveState"></a>}}
\if{latex}{\out{\hypertarget{method-ReactivePersistentObjectSet-retrieveState}{}}}
\subsection{Method \code{retrieveState()}}{
Load objects
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactivePersistentObjectSet$retrieveState(
  stateFiles = NULL,
  initTransaction = FALSE,
  invalidate = TRUE,
  includeChildren = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{stateFiles}}{list of state file paths}

\item{\code{initTransaction}}{boolean to init transaction}

\item{\code{invalidate}}{boolean to invalidate object}

\item{\code{includeChildren}}{boolean to include child nodes}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactivePersistentObjectSet-summary"></a>}}
\if{latex}{\out{\hypertarget{method-ReactivePersistentObjectSet-summary}{}}}
\subsection{Method \code{summary()}}{
Overview dataframe
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactivePersistentObjectSet$summary(
  fields = NULL,
  withSelf = TRUE,
  trimAttr = TRUE,
  uIDs = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fields}}{list of character field names}

\item{\code{withSelf}}{boolean to include self}

\item{\code{trimAttr}}{boolean to trim attribute names}

\item{\code{uIDs}}{list of character unique IDs}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactivePersistentObjectSet-runTasks"></a>}}
\if{latex}{\out{\hypertarget{method-ReactivePersistentObjectSet-runTasks}{}}}
\subsection{Method \code{runTasks()}}{
Run tasks for children
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactivePersistentObjectSet$runTasks(uIDs = NULL, mc.cores = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{uIDs}}{list of character of unique IDs}

\item{\code{...}}{ReactivePersistentObject$runTask}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactivePersistentObjectSet-cciaObjects"></a>}}
\if{latex}{\out{\hypertarget{method-ReactivePersistentObjectSet-cciaObjects}{}}}
\subsection{Method \code{cciaObjects()}}{
All objects
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactivePersistentObjectSet$cciaObjects(uIDs = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{uIDs}}{list of character of unique IDs}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactivePersistentObjectSet-cciaObjectsByUIDs"></a>}}
\if{latex}{\out{\hypertarget{method-ReactivePersistentObjectSet-cciaObjectsByUIDs}{}}}
\subsection{Method \code{cciaObjectsByUIDs()}}{
Objects by uID
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactivePersistentObjectSet$cciaObjectsByUIDs(uIDs)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{uIDs}}{list of character of unique IDs}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactivePersistentObjectSet-cciaObjectByUID"></a>}}
\if{latex}{\out{\hypertarget{method-ReactivePersistentObjectSet-cciaObjectByUID}{}}}
\subsection{Method \code{cciaObjectByUID()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactivePersistentObjectSet$cciaObjectByUID(uID)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactivePersistentObjectSet-cciaObjectsByAttr"></a>}}
\if{latex}{\out{\hypertarget{method-ReactivePersistentObjectSet-cciaObjectsByAttr}{}}}
\subsection{Method \code{cciaObjectsByAttr()}}{
Objects by attribute
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactivePersistentObjectSet$cciaObjectsByAttr(attrName, attrValue)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{attrName}}{character of attribute name}

\item{\code{attrValue}}{character/numeric of attribute value}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactivePersistentObjectSet-setCciaObjectUIDs"></a>}}
\if{latex}{\out{\hypertarget{method-ReactivePersistentObjectSet-setCciaObjectUIDs}{}}}
\subsection{Method \code{setCciaObjectUIDs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactivePersistentObjectSet$setCciaObjectUIDs(
  uIDs,
  invalidate = TRUE,
  reset = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactivePersistentObjectSet-getCciaObjectUIDs"></a>}}
\if{latex}{\out{\hypertarget{method-ReactivePersistentObjectSet-getCciaObjectUIDs}{}}}
\subsection{Method \code{getCciaObjectUIDs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactivePersistentObjectSet$getCciaObjectUIDs()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactivePersistentObjectSet-cciaObjectUIDs"></a>}}
\if{latex}{\out{\hypertarget{method-ReactivePersistentObjectSet-cciaObjectUIDs}{}}}
\subsection{Method \code{cciaObjectUIDs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactivePersistentObjectSet$cciaObjectUIDs(cciaObjects = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactivePersistentObjectSet-cciaObjectNames"></a>}}
\if{latex}{\out{\hypertarget{method-ReactivePersistentObjectSet-cciaObjectNames}{}}}
\subsection{Method \code{cciaObjectNames()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactivePersistentObjectSet$cciaObjectNames(cciaObjects = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactivePersistentObjectSet-cciaObjectMetas"></a>}}
\if{latex}{\out{\hypertarget{method-ReactivePersistentObjectSet-cciaObjectMetas}{}}}
\subsection{Method \code{cciaObjectMetas()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactivePersistentObjectSet$cciaObjectMetas(cciaObjects = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactivePersistentObjectSet-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ReactivePersistentObjectSet-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactivePersistentObjectSet$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
